### 2021-05-23 #3309 @whedon
[PRE REVIEW]: nnde: A Python package for solving differential equations using neural networks
**Submitting author:** @elwinter (<a href="http://orcid.org/0000-0001-5226-2107">Eric Winter</a>)
**Repository:** <a href="https://github.com/elwinter/nnde" target ="_blank">https://github.com/elwinter/nnde</a>
**Version:** 0.0.7
