### 2021-05-23 #3309 @arfon
:wave: @elwinter – thanks for your submission to JOSS. It looks like you're missing the BibTeX file to go with the `paper.md` file here?
