### 2021-05-23 #3309 @elwinter
DOH! Will upload it when i get home.

I hope all is well with you Arfon. I left ST in December for APL.

Eric

Get Outlook for Android<https://aka.ms/ghei36>
