### 2021-05-24 #3309 @arfon
> I hope all is well with you Arfon. I left ST in December for APL.

:wave: all good here thanks! Congrats on the new role, I hope APL is treating you well!
