### 2021-05-24 #3309 @elwinter
Hi Arfon. I added the paper.bib file to the nnde repo. Sorry about that.

Take care,
Eric

On Mon, May 24, 2021 at 4:30 AM Arfon Smith ***@***.***>
wrote:

> I hope all is well with you Arfon. I left ST in December for APL.
>
> 👋 all good here thanks! Congrats on the new role, I hope APL is treating
> you well!
>
