### 2021-06-06 3900a8a
Added recording of loss function history.
