### 2021-06-16 #3309 @diehlpk
Hey @rougier @betatim @pragyansmita @stsievert @road2stat @nilesh-patil would you be interested in reviewing this paper?
