### 2021-06-16 #3309 @stsievert
Sorry, I don't have the bandwidth right now. After a brief glance, I would encourage the author to look carefully at [the JOSS review criteria][1], especially [the documentation section][2].

[1]:https://joss.readthedocs.io/en/latest/review_criteria.html
[2]:https://joss.readthedocs.io/en/latest/review_criteria.html#documentation
