### 2021-06-16 #3309 @elwinter
Ok, I'll review it again, thanks.
Eric

Get Outlook for Android<https://aka.ms/ghei36>
