### 2021-06-22 #3309 @davismcc
Apologies, this application (solving differential equations) is too far outside my expertise to be confident evaluating it.

Get Outlook for iOS<https://aka.ms/o0ukef>
