### 2021-07-09 #3309 @elwinter
Hmm, OK. I’ll see what I (or my advisor) can come up with.

Take care

Eric

From: Patrick Diehl ***@***.***>
Sent: Thursday, July 8, 2021 11:48 AM
Cc: Eric Winter ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: nnde: A Python package for solving differential equations using neural networks (#3309)

Hi Patrick. Sorry for the delayed reply. I had suggested one reviewer from the JOSS list when I submitted the paper, but I see in my email archive that you are not having much luck getting a reviewer from your usual team. The only person on your team I could suggest is Arfon, but we know each other from working together at STScI, so that would not be appropriate, unless the reviews are blind. But even then, there’s a good chance he’d know it’s me, so Arfon should not be a candidate. Should I go through the list of your reviewers from the web site again? Or do you need to find an external reviewer at this point? Take care Eric From: Patrick Diehl @.> Sent: Tuesday, July 6, 2021 5:07 PM To: openjournals/joss-reviews @.> Cc: Eric Winter @.>; Mention @.> Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: nnde: A Python package for solving differential equations using neural networks (#3309 <https://github.com/openjournals/joss-reviews/issues/3309> ) @elwinter <https://github.com/elwinter>  https://github.com/elwinter Do you have some reviewers in mind? @elwinter <https://github.com/elwinter>  https://github.com/elwinter Do you have reviewers in mind? I would appreciated to get some names from you? — You are receiving this because you were mentioned. Reply to this email directly, view it on GitHub <#3309 (comment) <https://github.com/openjournals/joss-reviews/issues/3309#issuecomment-875081672> > , or unsubscribe https://github.com/notifications/unsubscribe-auth/ABSFWC6MKC7O57QUGMQHQ3TTWNV6PANCNFSM45LZJJXA . https://github.com/notifications/beacon/ABSFWC26FRGIPEAQH5JPRGLTWNV6PA5CNFSM45LZJJXKYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOGQUK7SA.gif

Yeah, I think I asked some of the reviewers in the JOSS list. I can wait for some time and ask the next round. I was thinking you have can recommend some external ones.m since you work on this topic.
