### 2021-07-09 #3465 @whedon
[REVIEW]: nnde: A Python package for solving differential equations using neural networks
**Submitting author:** @elwinter (<a href="http://orcid.org/0000-0001-5226-2107">Eric Winter</a>)
**Repository:** <a href="https://github.com/elwinter/nnde" target ="_blank">https://github.com/elwinter/nnde</a>
**Version:** v1.0
