### 2021-07-13 elwinter/nnde#4 @hayesall
Installation: Required packages in setup.py
Prerequisite packages are missing in `setup.py`, leading to errors such as the following:

```python
>>> from nnde.neuralnetwork.nnode1ivp import NNODE1IVPv
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/hayesall/Review/nnde/nnde/neuralnetwork/nnode1ivp.py", line 44, in <module>
    import numpy as np
ModuleNotFoundError: No module named 'numpy'
```

**Recommendations**:

- This should be fixed by setting `install_requires` in the `setup.py` (see [Python's documentation on distributing packages](https://packaging.python.org/guides/distributing-packages-using-setuptools/#install-requires)).
- Also see StackOverflow: [Good Practices for install_requires](https://stackoverflow.com/questions/65326080/python-setup-config-install-requires-good-practices/65326220#65326220)

---

**Reproduce with**:

```bash
# Clone repository
git clone git@github.com:elwinter/nnde.git
cd nnde

# Setup a fresh conda environment
conda create -n nnde python=3.7
conda activate nnde
pip install -e .

# Reproduce error
python
>>> from nnde.neuralnetwork.nnode1ivp import NNODE1IVPv

```
