### 2021-07-13 elwinter/nnde#5 @hayesall
Paper: Missing figure in paper.md?
Line 51 of the PDF ([download](https://raw.githubusercontent.com/openjournals/joss-papers/joss.03465/joss.03465/10.21105.joss.03465.pdf)) has a space followed by a period:

![Screenshot of three lines of text from the PDF. The final line has a broken link followed by a space and period](https://user-images.githubusercontent.com/11916674/125523236-7829300c-0bee-4969-8fa6-8d589a285da7.png)

It looks like this is either a broken link, or a missing figure. The raw markdown of `paper.md` refers to a file: `figures/Y_e.png`

```markdown
The `nnde` package was used to create a neural network with a single hidden layer and 10 hidden nodes and trained to solve this problem. The error in the trained solution for the case of $D=0.1$ is shown as a function of time in \autoref{fig:diff2d_error}.

![Difference between the trained neural network solution $\psi_t(x,y,t)$ and the analytical solution $\psi_a(x,y,t)$ of the diffusion problem in 2 spatial dimensions using `nnde` with 10 nodes.\label{fig:diff2d_error}](figures/Y_e.png)

```
