### 2021-07-13 elwinter/nnde#6 @hayesall
Documentation: Installation
Installation instructions should be included in the main README and associated documentation.

**Recommendation**:

- Command for installing from source on GitHub (e.g. clone and `pip install -e .`)
- Command for installing from PyPi (https://pypi.org/project/nnde/)
