### 2021-07-13 elwinter/nnde#6 @elwinter
Sorry, guess i assumed knowledge of pip. Will address this and prerequisites tonight or tomorrow and resubmit.

Thanks for the feedback
Eric

Get Outlook for Android<https://aka.ms/ghei36>
