### 2021-07-13 elwinter/nnde#7 @hayesall
Documentation: Statement of Need and Motivation for the software
The main README currently mentions the paper this implements.

**Recommendation**: Introduce the problem in 1-2 sentences, and suggest who the audience for this software is.

- ML researchers who want to use the method as a baseline?
- Researchers from general engineering fields who want to solve a differential equation?
