### 2021-07-13 elwinter/nnde#8 @hayesall
Installation Instructions: Demos
The [`nnde_demos`](https://github.com/elwinter/nnde_demos) repository has additional requirements beyond those required by the main repository.

**Recommendation**: A `requirements.txt` is a common convention where these are listed out. (`numpy`, `matplotlib`, `scipy`)
