### 2021-07-13 elwinter/nnde#9 @hayesall
Example Usage: Narrative examples to guide new users
The [`nnde_demos`](https://github.com/elwinter/nnde_demos) repository has good notes, but their presentation makes them difficult to follow.

**Recommendation**: Include Sphinx documentation or Jupyter notebooks to show steps of the computation alongside the narration for what occurs.

---

[`sphinx`](https://www.sphinx-doc.org/en/master/) documentation with the [`sphinx-gallery`](https://sphinx-gallery.github.io/stable/index.html) plugin is one option, or Jupyter notebooks can usually be rendered on GitHub.

For a real-world example of the `sphinx-gallery` approach, [here is a web page](https://imbalanced-learn.org/stable/auto_examples/api/plot_sampling_strategy_usage.html) that is generated from a [markdown document](https://github.com/scikit-learn-contrib/imbalanced-learn/blob/master/examples/api/plot_sampling_strategy_usage.py)
