### 2021-07-13 elwinter/nnde#10 @hayesall
Functionality Documentation: High-level software overview
Module and method documentation looks really good (e.g. all of the core classes and methods are documented in [`nnode1ivp`](https://github.com/elwinter/nnde/blob/master/nnde/neuralnetwork/nnode1ivp.py) and other methods).

At the moment it is difficult to find a starting point for:

1. what the core classes are
2. how to parameterize and use the core classes

**Recommendation**: [`sphinx`](https://www.sphinx-doc.org/en/master/) documentation would be ideal. At an absolute minimum, a document should list the core classes/methods, an example for how to parameterize them, and what some of the inputs/outputs should look like.
