### 2021-07-13 elwinter/nnde#11 @hayesall
Automated Tests: How does a user run them?
**Recommendation**: Document how tests should be ran (briefly in the README, on a webpage, or in a [`CONTRIBUTING.md` guide](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors)).

---

`pytest` seems to pick most of them up:

```
# Clone
git clone git@github.com:elwinter/nnde.git
cd nnde

# Install and run
pip install pytest
pytest
```
