### 2021-07-13 elwinter/nnde#12 @hayesall
Automated Tests: Continuous Integration and Testing
**Recommendation**: Run tests automatically on a push. [GitHub Actions](https://docs.github.com/en/actions/guides/building-and-testing-python) or a Continuous Integration (CI) service should help here (Circle-CI, Travis-CI, AppVeyor, etc.)
