### 2021-07-13 elwinter/nnde#14 @hayesall
Community Guidelines: How does a user contribute?
**Recommendation**: Add `CONTRIBUTING.md` with brief notes for how to contribute:

1. What contributions would be helpful (bug reports, patches, documentation)
2. Is there a formal review process, etc.

GitHub suggests placing these in a [`CONTRIBUTING.md`](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) file, which GitHub looks for in a `.github/` directory at the base of a repository.
