### 2021-07-13 elwinter/nnde#15 @hayesall
Community Guidelines: How to report bugs and ask questions?
**Recommendation**: Add `CONTRIBUTING.md` note for how to report issues or ask questions. These might also be mentioned in a [`.github/ISSUE_TEMPLATE`](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository) file.
