### 2021-07-13 elwinter/nnde#16 @hayesall
Paper Feedback: Analytic methods for derivative calculations
The "Statement of need" (lines 19-22) mentioned that existing deep learning frameworks can speed up computation, but **cannot** perform analytical derivatives needed for the trial function method implemented by `nnde`.

![Screenshot of the paper, where it shows something similar to the text summarized above.](https://user-images.githubusercontent.com/11916674/125535312-75146622-1797-40e0-b50a-0b63a3a49f52.png)

Can this point be clarified or expanded?

For example: PyTorch implements methods for automatic differentiation ([autograd](https://pytorch.org/tutorials/beginner/blitz/autograd_tutorial.html)) and numerical methods to compute gradients. What is an example where these are insufficient?
