### 2021-07-13 elwinter/nnde#6 @rweigel
It is standard to include dependencies in `setup.py`. Without it, the user would need to search for all of the import statements to determine dependencies.
