### 2021-07-13 elwinter/nnde#6 @rweigel
Test by creating a virtual environment and executing `pip install nnde` followed by execution of a test or demo. I thought that I did this when testing but clearly I missed it.
