### 2021-07-13 elwinter/nnde#6 @rweigel
I suggest adding the following section to `README.md`

# Installation

To install the library, use

```bash
git clone https://github.com/elwinter/nnde.git
cd nnde; pip install -e .
```

or

```bash
pip install nnde
```

The demos are stored in a separate repository. To use them, execute

```
git clone https://github.com/elwinter/nnde_demos
cd nnde_demos
python lagaris01_demo.py
python nnpde2diff_demo.py
```
