### 2021-07-13 #3465 @hayesall
I made the first review pass @elwinter. I need some additional time to check functionality, but all of the points below should be possible to do parallel to that.

### Overall Feedback

All of the components are here (software, docstrings, tests), the final step is to make these components accessible to a user. Many of these could be addressed by incorporating a documentation system (e.g. [`sphinx`](https://www.sphinx-doc.org/en/master/)) to pull docstrings and collect high-level overviews.

_If possible_: After writing the high-level documentation; schedule a meeting with a colleague (someone who is fairly comfortable in Python and knows a little about this problem setting), give them links to the software and documentation, and watch them get started with it.

### Functionality / Installation

- [ ] https://github.com/elwinter/nnde/issues/4
- [ ] https://github.com/elwinter/nnde/issues/6
- [ ] https://github.com/elwinter/nnde/issues/8

### Documentation, Testing, and Community Guidelines

- [ ] https://github.com/elwinter/nnde/issues/7
- [ ] https://github.com/elwinter/nnde/issues/9
- [ ] https://github.com/elwinter/nnde/issues/10
- [ ] https://github.com/elwinter/nnde/issues/11
- [ ] https://github.com/elwinter/nnde/issues/12
- [ ] https://github.com/elwinter/nnde/issues/13
- [ ] https://github.com/elwinter/nnde/issues/14
- [ ] https://github.com/elwinter/nnde/issues/15

### Paper Feedback

- [ ] https://github.com/elwinter/nnde/issues/5
- [ ] https://github.com/elwinter/nnde/issues/16
