### 2021-07-13 #3465 @elwinter
Excellent. Thanks for your time. I'll start addressing these over the next week or so and resubmit.

Take care
Eric

Get Outlook for Android<https://aka.ms/ghei36>
