### 2021-07-13 elwinter/nnde#6 @hayesall
Good recommendation @rweigel. #4 also includes the point for adding all requirements under `install_requires` in `setup.py`.
