### 2021-07-17 ff03557
Added simple_01 example equation.
