### 2021-07-21 f27bb97
Added figures directory for paper.
