+++ setup.py
@@ -23,5 +23,5 @@ setuptools.setup(
     python_requires='>=3.6',
     tests_require=['nose', 'unittest'],
     test_suite='nose.collector',
-    install_requires=['importlib', 'math', 'numpy', 'scipy']
+    install_requires=['numpy', 'scipy']
 )
