### 2021-07-21 c85fab5
remove standard libs
