### 2021-07-24 88ed2a2
Added matplotlib to required install packages.
