+++ tox.ini
@@ -1,0 +1,15 @@
+# tox (https://tox.readthedocs.io/) is a tool for running tests
+# in multiple virtualenvs. This configuration file will run the
+# test suite on all supported python versions. To use it, "pip install tox"
+# and then run "tox" from this directory.
+
+[tox]
+envlist = py38
+
+[testenv]
+deps =
+    pytest
+    numpy
+    scipy
+commands =
+    pytest
