### 2021-07-28 elwinter/nnde#16 @byekrang
Would you please explain about extra functionalities of nnde with this version of [Autograd](https://github.com/HIPS/autograd) (not the PyTorch's one) and [JAX](https://github.com/google/jax) considering [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)?
