+++ nnde/differentialequation/differentialequation.py
@@ -1,19 +1,41 @@
-"""Abstract base class for differential equations
+"""Abstract base class for differential equations.
 
-This module defines the methods required for all differential equation
-objects used in the nnde software.
+Deprecation Warning
+-------------------
+None
 
-This class is an abstract base class. It must be subclassed to be useful. All
-methods in this class will raise NNDEException if called.
+Extended Summary
+----------------
+This module defines the DifferentialEquation, which is an abstract
+class defining methods by for all differential equation classes used
+in the nnde package.
 
-Attributes:
-    None
+This class is an abstract base class. It must be subclassed to be
+useful. All methods in this class will raise NNDEException if called.
 
-Methods:
-    __init__() - Constructor
-    G() - Function for differential equation, in the form G() = 0,
-    where G() is a function of the independent variables x, the solution
-    Y(x), and its derivatives.
+Classes
+-------
+DifferentialEquation
+
+Functions
+---------
+None
+
+See Also
+--------
+
+Notes
+-----
+
+References
+----------
+
+Examples
+--------
+
+Authors
+-------
+Eric Winter (eric.winter62@gmail.com)
 """
 
 
@@ -21,7 +43,57 @@ from nnde.exceptions.nndeexception import NNDEException
 
 
 class DifferentialEquation:
-    """Abstract base class for differential equation objects"""
+    """Abstract base class for differential equation objects.
+
+    Deprecation Warning
+    -------------------
+    None
+
+    Extended Summary
+    ----------------
+    This abstract base class defines the required methods for all
+    differential equation classes in the nnde package.
+
+    Parameters
+    ----------
+
+    Attributes
+    ----------
+
+    Returns
+    -------
+
+    Yields
+    ------
+
+    Receives
+    --------
+
+    Other Parameters
+    ----------------
+    Infrequently-used parameters go here.
+
+    Raises
+    ------
+
+    Warns
+    -----
+
+    Warnings
+    --------
+
+    See Also
+    --------
+
+    Notes
+    -----
+
+    References
+    ----------
+
+    Examples
+    --------
+    """
 
     def __init__(self, *args, **kwargs):
         """Constructor for DifferentialEquation objects"""
