### 2021-07-31 elwinter/nnde#4 @elwinter
This has been addressed in the latest setup.py as follows:

    install_requires=['matplotlib', 'numpy', 'scipy']

It was verified in a bare-bones conda environment created for testing.
