### 2021-07-31 elwinter/nnde#6 @elwinter
Instructions for installing from PyPI using pip, and by cloning the GitHub repository, have been added to the top-level README.md.
