### 2021-07-31 elwinter/nnde#8 @elwinter
A requirements.txt file listing numpy, scipy, and matplotlib has been added, and tested to ensure that it can be used to install the required modules with:

pip install -r requirements.txt
