### 2021-08-01 elwinter/nnde#9 @elwinter
I added a Jupyter notebook (tutorial.ipynb) to the nnde_demos repository (https://github.com/elwinter/nnde_demos) that walks the user through the process of solving a 1st-order ODE IVP using the nnde ppackage and examining the resulting network and the solutions it generates.
