### 2021-08-07 875e5fc
Started updating docstrings.
