### 2021-08-07 c1e282d
Added software overview to README.md, and added examples directory with tutorial.ipynb.
