### 2021-08-07 elwinter/nnde#10 @elwinter
I added a section in README.md that provides an overview of the structure of the package and discusses the major classes. I also added an examples directory at the top level, which contains a Jupyter notebook (tutorial.ipynb) that provides a detailed walkthrough of how to use the nnde package to solve a differential equation.
