### 2021-08-07 elwinter/nnde#11 @elwinter
README.md was modified to illustrate running the test code with "python setup.py test" and "pytest".
