### 2021-08-07 0123d7c
Modified testing section, added discussion of big reports and contributions.
