### 2021-08-07 elwinter/nnde#14 @elwinter
README.md was updated with instructions for providing bug reports and code contributions.
