### 2021-08-07 a6347e0
Added CONTRIBUTING.md.
