+++ CONTRIBUTING.md
@@ -1,0 +1,9 @@
+# Contributing to nnde
+
+If you have modified the nnde package for your own work, and wish to contribute to the project, we welcome your contributions of code and documentation. To contribute, fork the repository on GitHub, and submit a pull request at https://github.com/elwinter/nnde.
+
+# Contact
+
+Submit bug reports and feature requests on the [repository issue tracker](https://github.com/elwinter/nnde/issues).
+
+Eric Winter <eric.winter62@gmail.com>
