### 2021-08-07 elwinter/nnde#15 @elwinter
Added a CONTRIBUTING.md file with instructions.
