### 2021-08-11 #3465 @elwinter
Hi Patrick. Revisions are proceeding. I have closed all but 3 issues opened by the reviewers now, and I hope to have the rest closed by the end of this week.

Take care

Eric

From: Patrick Diehl ***@***.***>
Sent: Monday, August 9, 2021 10:14 AM
Cc: Eric Winter ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: nnde: A Python package for solving differential equations using neural networks (#3465)

Hey @elwinter <https://github.com/elwinter>  how is your revision going?
