### 2021-08-14 733066b
Removed demo github actions file. Renamed remaining file.
