### 2021-08-14 16b8292
Added code to install pytest.
