+++ .github/workflows/github-run-nnde-tests.yml
@@ -11,7 +11,10 @@ jobs:
         uses: actions/checkout@v2
       - run: echo "💡 The ${{ github.repository }} repository has been cloned to the runner."
       - run: echo "🖥️ The workflow is now ready to test your code on the runner."
-      - name: List files in the repository
-        run: |
+      - name: Install pytest to run the tests.
+        run:
+          pip install pytest
+      - name: Run the tests.
+        run:
           pytest
       - run: echo "🍏 This job's status is ${{ job.status }}."
