### 2021-08-14 8ede253
Added code to install numpy and scipy.
