+++ .github/workflows/github-run-nnde-tests.yml
@@ -13,7 +13,7 @@ jobs:
       - run: echo "🖥️ The workflow is now ready to test your code on the runner."
       - name: Install pytest to run the tests.
         run:
-          pip install pytest
+          pip install pytest numpy scipy
       - name: Run the tests.
         run:
           pytest
