### 2021-08-14 1eaf950
Added code to run tox for tests.
