### 2021-08-14 elwinter/nnde#12 @elwinter
I configured a GitHub action to automatically run the test suite using pytest and tox. The runs are triggered on a repository push. I'll remove one of the tests runs once I decide which is more useful, report-wise.
