+++ nnde/differentialequation/examples/tests/test_diff1d_half.py
@@ -1,112 +1,364 @@
+"""Tests for the diff1d_half module."""
+
+
+import unittest
+
 import numpy as np
 
 from nnde.differentialequation.examples.diff1d_half import (
     C, D,
-    G, dG_dY, dG_ddelY, dG_ddel2Y,
-    bc, f0, f1, Y0, delbc, del2bc,
+    G,
+    f0, f1, Y0, bc,
+    df0_dx, df0_dt, df1_dx, df1_dt, dY0_dx, dY0_dt, delbc,
+    d2f0_dx2, d2f0_dt2, d2f1_dx2, d2f1_dt2, d2Y0_dx2, d2Y0_dt2, del2bc,
+    dG_dY, dG_dY_dx, dG_dY_dt, dG_ddelY,
+    dG_d2Y_dx2, dG_d2Y_dt2, dG_ddel2Y,
     A, delA, del2A,
-    Ya, delYa, del2Ya
+    Ya, dYa_dx, dYa_dt, delYa, d2Ya_dx2, d2Ya_dt2, del2Ya
 )
 
 
+# Grid points for testing.
+xx = np.linspace(0, 1, 11)
+tt = np.linspace(0, 1, 11)
+
+
+class TestBuilder(unittest.TestCase):
+    """Tests for the diff1d_half module."""
+
+    def test_G(self):
+        """Test the differential equation."""
+        xt = (0, 0)
+        Yt = 0
+        delYt = (0, 0)
+        del2Yt = (0, 0)
+        G_ref = 0
+        self.assertAlmostEqual(G(xt, Yt, delYt, del2Yt), G_ref)
+        xt = (1, 1)
+        Yt = 2
+        delYt = (3, 4)
+        del2Yt = (5, 6)
+        G_ref = delYt[1] - D*del2Yt[0]
+        self.assertAlmostEqual(G(xt, Yt, delYt, del2Yt), G_ref)
+
+    def test_f0(self):
+        """Test boundary condition at (x, t) = (0, t)."""
+        f0_ref = C
+        for x in xx:
+            for t in tt:
+                xt = (x, t)  # x is ignored
+                self.assertAlmostEqual(f0(xt), f0_ref)
+
+    def test_f1(self):
+        """Test boundary condition at (x, t) = (1, t)."""
+        f1_ref = C
+        for x in xx:
+            for t in tt:
+                xt = (x, t)  # x is ignored
+                self.assertAlmostEqual(f1(xt), f1_ref)
+
+    def test_Y0(self):
+        """Test boundary condition at (x, t) = (x, 0)."""
+        Y0_ref = C
+        for x in xx:
+            for t in tt:
+                xt = (x, t)  # t is ignored
+                self.assertAlmostEqual(Y0(xt), Y0_ref)
+
+    def test_bc(self):
+        """Test that the boundary conditions are continuous at corners."""
+        xt = (0, 0)
+        self.assertAlmostEqual(bc[0][0](xt), bc[1][0](xt))
+        xt = (1, 0)
+        self.assertAlmostEqual(bc[0][1](xt), bc[1][0](xt))
+
+    def test_df0_dx(self):
+        """Test the first derivative of f0 wrt x."""
+        df0_dx_ref = 0
+        for x in xx:
+            for t in tt:
+                xt = (x, t)  # x is ignored
+                self.assertAlmostEqual(df0_dx(xt), df0_dx_ref)
+
+    def test_df0_dt(self):
+        """Test the first derivative of f0 wrt t."""
+        df0_dt_ref = 0
+        for x in xx:
+            for t in tt:
+                xt = (x, t)  # x is ignored
+                self.assertAlmostEqual(df0_dt(xt), df0_dt_ref)
+
[…]
