+++ nnde/differentialequation/examples/diff1d_half.py
@@ -3,7 +3,7 @@
 This module implements a 1-D diffusion PDE.
 
 Note that an upper-case 'Y' is used to represent the Greek psi, which
-represents the problem solution Y(x,t). Y(x, t) is normalized to the
+represents the problem solution Y(x, t). Y(x, t) is normalized to the
 range [0, 1], and is treated as unitless.
 
 The equation is defined on the domain:
@@ -45,9 +45,6 @@ Eric Winter (eric.winter62@gmail.com)
 """
 
 
-import numpy as np
-
-
 # Diffusion coefficient (L**2/T, where L is a length, and T is a time).
 D = 0.1
 
@@ -56,8 +53,8 @@ C = 0.5
 
 
 def G(xt, Y, delY, del2Y):
-    """The differential equation in standard form.
-    
+    """Compute the differential equation in standard form.
+
     Compute the value of the differential equation in standard form. For a
     perfect solution, the value should be 0.
 
@@ -71,28 +68,27 @@ def G(xt, Y, delY, del2Y):
         Values for dY/dx and dY/dt, in that order.
     del2Y : array-like of 2 float
         Values for d2Y/dx2 and d2Y/dt2, in that order.
-    
+
     Returns
     -------
     result : float
         Value of the differential equation.
     """
-    (x, t) = xt
-    (dY_dx, dY_dt) = delY
-    (d2Y_dx2, d2Y_dt2) = del2Y
+    dY_dt = delY[1]
+    d2Y_dx2 = del2Y[0]
     return dY_dt - D*d2Y_dx2
 
 
 def f0(xt):
     """Boundary condition at (x, t) = (0, t).
-    
+
     Compute the value of the solution at (x, t) = (0, t).
 
     Parameters
     ----------
     xt : array-like of 2 float
         Values for x and t, in that order.
-    
+
     Returns
     -------
     result : float
@@ -103,14 +99,14 @@ def f0(xt):
 
 def f1(xt):
     """Boundary condition at (x, t) = (1, t).
-    
+
     Compute the value of the solution at (x, t) = (1, t).
 
     Parameters
     ----------
     xt : array-like of 2 float
         Values for x and t, in that order.
-    
+
     Returns
     -------
     result : float
@@ -120,153 +116,530 @@ def f1(xt):
 
 
 def Y0(xt):
-    """Boundary condition at (x,t) = (x,0)"""
+    """Boundary condition at (x, t) = (x, 0).
+
+    Compute the value of the solution at (x, t) = (x, 0).
+
+    Parameters
+    ----------
+    xt : array-like of 2 float
+        Values for x and t, in that order.
+
+    Returns
+    -------
+    result : float
+        Value of the solution at (x, t) = (x, 0).
+    """
     return C
 
 
+# Gather the boundary condition functions in a single array.
 bc = [[f0, f1], [Y0, None]]
 
 
 def df0_dx(xt):
-    """1st derivative of BC wrt x at (x,t) = (0,t)"""
+    """1st derivative of BC wrt x at (x, t) = (0, t).
+
+    Compute the 1st derivative of BC wrt x at (x, t) = (0, t).
+
+    Parameters
+    ----------
+    xt : array-like of 2 float
+        Values for x and t, in that order.
+
+    Returns
+    -------
+    result : float
+        Value of the 1st derivative of BC wrt x at (x, t) = (0, t).
+    """
     return 0
 
 
 def df0_dt(xt):
-    """1st derivative of BC wrt t at (x,t) = (0,t)"""
+    """1st derivative of BC wrt t at (x, t) = (0, t).
+
+    Compute the 1st derivative of BC wrt t at (x, t) = (0, t).
+
+    Parameters
+    ----------
+    xt : array-like of 2 float
[…]
