### 2021-08-15 5149b9f
Corrected gradient and Laplacian tests, added comments about bad function names.
