### 2021-08-15 7987c52
Converted diff1d_halfsine_increase tests to unittest format, massive linting.
