+++ nnde/differentialequation/examples/diff1d_halfsine_increase.py
@@ -1,22 +1,29 @@
-"""
-This module implements a 1-D diffusion PDE
+"""The 1-D diffusion equation.
+
+This module implements a 1-D diffusion PDE.
 
 Note that an upper-case 'Y' is used to represent the Greek psi, which
-represents the problem solution Y(x,t).
+represents the problem solution Y(x,t). Y(x, t) is normalized to the
+range [0, 1], and is treated as unitless.
 
-The equation is defined on the domain (x, t) in [[0, 1],[0, ]].
+The equation is defined on the domain:
+
+  0 <= x <= 1
+  0 <= t
 
 The analytical form of the equation is:
 
-  G([x, t], Y, delY, del2Y) = dY_dt - D*d2Y_dx2 = 0
+  G(xt, Y, delY, del2Y) = dY_dt - D*d2Y_dx2 = 0
 
 where:
 
-[x, t] are the independent variables.
-delY is the vector (dY/dx, dY/dt)
-del2Y is the vector (d2Y/dx2, d2Y/dt2)
+xt is the vector (x, t).
+Y is the solution to be found.
+delY is the gradient vector (dY/dx, dY/dt)->(dY_dx, dY_dt).
+del2Y is the Laplacian component vector (d2Y/dx2, d2Y/dt2)->
+(d2Y_dx2, d2Y_dt2).
 
-With boundary conditions:
+The boundary conditions are:
 
 Y(0, t) = 0
 Y(1, t) = a*t
@@ -29,164 +36,504 @@ Ya(x, t) = a*t*x + sin(pi*x)*(cosh(pi**2*t*D) - sinh(pi**2*t*D))
            + 0.5*SUM(k=1, inf, 2*(-1)**k*a*(1 - exp(-pi**2*t*D*k**2))
                                *sin(pi*x*k)/(pi**3*k**3)
 
-Todo:
-    * Add function annotations.
-    * Add variable annotations.
+Authors
+-------
+Eric Winter (eric.winter62@gmail.com)
 """
 
 
 from math import exp, cos, cosh, pi, sin, sinh
-import numpy as np
 
 
-# Diffusion coefficient
+# Diffusion coefficient (L**2/T, where L is a length, and T is a time).
 D = 0.1
 
-# Boundary increase rate at x=1
+# Boundary increase rate at x=1 (1/T).
 a = 0.1
 
-# Number of terms in analytical summation
+# Number of terms in analytical solution summation.
 kmax = 800
 
 
 def G(xt, Y, delY, del2Y):
-    """The differential equation in standard form"""
-    (x, t) = xt
-    (dY_dx, dY_dt) = delY
-    (d2Y_dx2, d2Y_dt2) = del2Y
+    """Compute the differential equation in standard form.
+
+    Compute the value of the differential equation in standard form. For a
+    perfect solution, the value should be 0.
+
+    Parameters
+    ----------
+    xt : array-like of 2 float
+        Values for x and t, in that order.
+    Y : float
+        Current solution value.
+    delY : array-like of 2 float
+        Values for dY/dx and dY/dt, in that order.
+    del2Y : array-like of 2 float
+        Values for d2Y/dx2 and d2Y/dt2, in that order.
+
+    Returns
+    -------
+    result : float
+        Value of the differential equation.
+    """
+    dY_dt = delY[1]
+    d2Y_dx2 = del2Y[0]
     return dY_dt - D*d2Y_dx2
 
 
 def f0(xt):
-    """Boundary condition at (x,t) = (0,t)"""
+    """Boundary condition at (x, t) = (0, t).
+
+    Compute the value of the solution at (x, t) = (0, t).
+
+    Parameters
+    ----------
+    xt : array-like of 2 float
+        Values for x and t, in that order.
+
+    Returns
+    -------
+    result : float
[…]
