### 2021-08-16 40d8aab
Converted diff1d_halfsine tests to unittest format, and extensive linting.
