### 2021-08-16 d29544e
Changed diff1d_one tests to unittest format, extensive linting.
