+++ nnde/differentialequation/examples/diff1d_one.py
@@ -1,22 +1,29 @@
-"""
-This module implements a 1-D diffusion PDE
+"""The 1-D diffusion equation.
+
+This module implements a 1-D diffusion PDE.
 
 Note that an upper-case 'Y' is used to represent the Greek psi, which
-represents the problem solution Y(x,t).
+represents the problem solution Y(x, t). Y(x, t) is normalized to the
+range [0, 1], and is treated as unitless.
 
-The equation is defined on the domain (x,t) in [[0,1],[0,]].
+The equation is defined on the domain:
+
+  0 <= x <= 1
+  0 <= t
 
 The analytical form of the equation is:
 
-  G(x, Y, delY, del2Y) = dY_dt - D*d2Y_dx2 = 0
+  G(xt, Y, delY, del2Y) = dY_dt - D*d2Y_dx2 = 0
 
 where:
 
-xv is the vector (x,t)
-delY is the vector (dY/dx, dY/dt)
-del2Y is the vector (d2Y/dx2, d2Y/dt2)
+xt is the vector (x, t).
+Y is the solution to be found.
+delY is the gradient vector (dY/dx, dY/dt)->(dY_dx, dY_dt).
+del2Y is the Laplacian component vector (d2Y/dx2, d2Y/dt2)->
+(d2Y_dx2, d2Y_dt2).
 
-With boundary conditions:
+The boundary conditions are:
 
 Y(0, t) = C = 1
 Y(1, t) = C = 1
@@ -27,24 +34,46 @@ conditions:
 
 Ya(x, t) = 1
 
-Todo:
-    * Add function annotations.
-    * Add variable annotations.
+Note
+----
+This simple equation has a constant as a solution, and was used to ensure
+solution stability.
+
+Authors
+-------
+Eric Winter (eric.winter62@gmail.com)
 """
 
 
-import numpy as np
-
-
-# Diffusion coefficient
+# Diffusion coefficient (L**2/T, where L is a length, and T is a time).
 D = 0.1
 
-# Constant value of profile
+# Initial value of profile (unitless).
 C = 1
 
 
 def G(xt, Y, delY, del2Y):
-    """The differential equation in standard form"""
+    """Compute the differential equation in standard form.
+
+    Compute the value of the differential equation in standard form. For a
+    perfect solution, the value should be 0.
+
+    Parameters
+    ----------
+    xt : array-like of 2 float
+        Values for x and t, in that order.
+    Y : float
+        Current solution value.
+    delY : array-like of 2 float
+        Values for dY/dx and dY/dt, in that order.
+    del2Y : array-like of 2 float
+        Values for d2Y/dx2 and d2Y/dt2, in that order.
+
+    Returns
+    -------
+    result : float
+        Value of the differential equation.
+    """
     (x, t) = xt
     (dY_dx, dY_dt) = delY
     (d2Y_dx2, d2Y_dt2) = del2Y
@@ -52,163 +81,570 @@ def G(xt, Y, delY, del2Y):
 
 
 def f0(xt):
-    """Boundary condition at (x,t) = (0,t)"""
+    """Boundary condition at (x, t) = (0, t).
+
+    Compute the value of the solution at (x, t) = (0, t).
+
+    Parameters
+    ----------
+    xt : array-like of 2 float
+        Values for x and t, in that order.
+
+    Returns
+    -------
+    result : float
+        Value of the solution at (x, t) = (0, t).
+    """
     return C
 
 
 def f1(xt):
-    """Boundary condition at (x,t) = (1,t)"""
+    """Boundary condition at (x, t) = (1, t).
+
+    Compute the value of the solution at (x, t) = (1, t).
+
+    Parameters
+    ----------
+    xt : array-like of 2 float
+        Values for x and t, in that order.
+
+    Returns
+    -------
+    result : float
[…]
