### 2021-08-16 4532ae7
Converted diff1d_zero tests to unittest format, extensive linting.
