### 2021-08-16 c1fce32
Converted diff2d_halfsine_increase tests to unittest format, extensive linting.
