+++ nnde/differentialequation/examples/tests/test_diff2d_halfsine_increase.py
@@ -1,104 +1,570 @@
+"""Tests for the diff2d_halfsine_increase module."""
+
+
+from math import cos, pi, sin
+import unittest
+
 import numpy as np
 
 from nnde.differentialequation.examples.diff2d_halfsine_increase import (
     a, D,
-    G, dG_dY, dG_ddelY, dG_ddel2Y,
-    bc, f0, f1, g0, g1, Y0, delbc, del2bc,
+    G,
+    f0, f1, g0, g1, Y0, bc,
+    df0_dx, df0_dy, df0_dt, df1_dx, df1_dy, df1_dt,
+    dg0_dx, dg0_dy, dg0_dt, dg1_dx, dg1_dy, dg1_dt,
+    dY0_dx, dY0_dy, dY0_dt, delbc,
+    d2f0_dx2, d2f0_dy2, d2f0_dt2, d2f1_dx2, d2f1_dy2, d2f1_dt2,
+    d2g0_dx2, d2g0_dy2, d2g0_dt2, d2g1_dx2, d2g1_dy2, d2g1_dt2,
+    d2Y0_dx2, d2Y0_dy2, d2Y0_dt2, del2bc,
+    dG_ddY_dx, dG_ddY_dy, dG_ddY_dt, dG_ddelY,
+    dG_dd2Y_dx2, dG_dd2Y_dy2, dG_dd2Y_dt2, dG_ddel2Y,
     A, delA, del2A,
 )
 
 
+# Grid points for testing.
+xx = np.linspace(0, 1, 11)
+yy = np.linspace(0, 1, 11)
+tt = np.linspace(0, 1, 11)
+
+
+class TestBuilder(unittest.TestCase):
+    """Tests for the diff2d_halfsine_increase module."""
+
+    def test_G(self):
+        """Test the differential equation."""
+        Yt = 2
+        delYt = (3, 4, 5)
+        del2Yt = (5, 6, 7)
+        for x in xx:
+            for y in yy:
+                for t in tt:
+                    xyt = (x, y, t)
+                    G_ref = delYt[2] - D*(del2Yt[0] + del2Yt[1])
+                    self.assertAlmostEqual(G(xyt, Yt, delYt, del2Yt), G_ref)
+
+    def test_f0(self):
+        """Test boundary condition at (x, y, t) = (0, y, t)."""
+        f0_ref = 0  # For all inputs.
+        for x in xx:
+            for y in yy:
+                for t in tt:
+                    xyt = (x, y, t)
+                    self.assertAlmostEqual(f0(xyt), f0_ref)
+
+    def test_f1(self):
+        """Test boundary condition at (x, y, t) = (1, y, t)."""
+        for x in xx:
+            for y in yy:
+                for t in tt:
+                    xyt = (x, y, t)
+                    f1_ref = a*t*sin(pi*y)
+                    self.assertAlmostEqual(f1(xyt), f1_ref)
+
+    def test_g0(self):
+        """Test boundary condition at (x, y, t) = (x, 0, t)."""
+        g0_ref = 0  # For all inputs.
+        for x in xx:
+            for y in yy:
+                for t in tt:
+                    xyt = (x, y, t)
+                    self.assertAlmostEqual(g0(xyt), g0_ref)
+
+    def test_g1(self):
+        """Test boundary condition at (x, y, t) = (x, 1, t)."""
+        g1_ref = 0  # For all inputs.
+        for x in xx:
+            for y in yy:
+                for t in tt:
+                    xyt = (x, y, t)
+                    self.assertAlmostEqual(g1(xyt), g1_ref)
+
+    def test_Y0(self):
+        """Test boundary condition at (x, y, t) = (x, y, 0)."""
+        for x in xx:
[…]
