### 2021-08-16 72fe797
Converted diff2d_halfsine tests to unittest format, and extensive linting.
