+++ nnde/differentialequation/examples/diff2d_halfsine.py
@@ -1,52 +1,76 @@
-"""
-This module implements a 2-D diffusion PDE.
+"""The 1-D diffusion equation.
+
+This module implements a 12-D diffusion PDE.
 
 Note that an upper-case 'Y' is used to represent the Greek psi, which
-represents the problem solution Y(x, y, t).
+represents the problem solution Y(x, y, t). Y(x, y, t) is normalized to the
+range [0, 1], and is treated as unitless.
 
-The equation is defined on the domain (x, y, t) in
-([0, 1], [0, 1], [0, inf]).
+The equation is defined on the domain:
+
+  0 <= x <= 1
+  0 <= y <= 1
+  0 <= t
 
 The analytical form of the equation is:
 
-  G([x, y, t], Y, delY, del2Y) = dY_dt - D*(d2Y_dx2 + d2Y_dy2) = 0
+  G(xyt, Y, delY, del2Y) = dY_dt - D*(d2Y_dx2 + d2Y_dy2) = 0
 
 where:
 
-x, y, t are the independent variables
-delY is the vector (dY/dx, dY/dy, dY/dt)
-del2Y is the vector (d2Y/dx2, d2Y_dy2, d2Y/dt2)
-D is the diffusion coefficient
+xyt is the vector (x, y, t).
+Y is the solution to be found.
+delY is the gradient vector (dY/dx, dY/dy, dY/dt).
+del2Y is the Laplacian component vector (d2Y/dx2, d2Y/dy2, d2Y/dt2).
 
-With boundary conditions:
+The boundary conditions are:
 
 Y(0, y, t) = 0
 Y(1, y, t) = 0
 Y(x, 0, t) = 0
 Y(x, 1, t) = 0
-Y(x, 0) = sin(pi*x)*sin(pi*y)
+Y(x, y, 0) = sin(pi*x)*sin(pi*y)
 
 This equation has the analytical solution for the supplied initial
 conditions:
 
 Ya(x, y, t) = exp(-2*pi**2*D*t)*sin(pi*x)*sin(pi*y)
 
-Todo:
-    * Add function annotations.
-    * Add variable annotations.
+Authors
+-------
+Eric Winter (eric.winter62@gmail.com)
 """
 
 
 from math import cos, exp, pi, sin
-import numpy as np
 
 
-# Diffusion coefficient
+# Diffusion coefficient (L**2/T, where L is a length, and T is a time).
 D = 0.1
 
 
 def G(xyt, Y, delY, del2Y):
-    """The differential equation in standard form"""
+    """Compute the differential equation in standard form.
+
+    Compute the value of the differential equation in standard form. For a
+    perfect solution, the value should be 0.
+
+    Parameters
+    ----------
+    xyt : array-like of 3 float
+        Values for x, y, and t, in that order.
+    Y : float
+        Current solution value.
+    delY : array-like of 3 float
+        Values for dY/dx, dY/dy, and dY/dt, in that order.
+    del2Y : array-like of 3 float
+        Values for d2Y/dx2, d2Y/dy2, and d2Y/dt2, in that order.
+
+    Returns
+    -------
+    result : float
+        Value of the differential equation.
+    """
     (x, y, t) = xyt
     (dY_dx, dY_dy, dY_dt) = delY
     (d2Y_dx2, d2Y_dy2, d2Y_dt2) = del2Y
@@ -54,248 +78,868 @@ def G(xyt, Y, delY, del2Y):
 
 
 def f0(xyt):
-    """Boundary condition at (x, y, t) = (0, y, t)"""
+    """Boundary condition at (x, y, t) = (0, y, t).
+
+    Compute the value of the solution at (x, y, t) = (0, y, t).
+
+    Parameters
+    ----------
+    xyt : array-like of 3 float
+        Values for x, y, and t, in that order.
+
[…]
