### 2021-08-21 elwinter/nnde#16 @elwinter
We updated the paper to address this issue by describing how the code was started before the widespread adoption of TensorFlow 2. We began to convert the software to use TensorFlow 2 several months ago, and early results are promising, so we wanted to document this particular stage in the project prior to further development.
