### 2021-08-22 8eeb2a7
Fixed docstring and test typos.
