### 2021-08-22 e3c7dd4
Fixed docstring typo.
