### 2021-08-22 a9693e1
Docstrings, and rewrote tests into unittest format.
