+++ nnde/differentialequation/examples/tests/test_diff3d_halfsine_increase.py
@@ -1,136 +1,916 @@
+"""Tests for the diff3d_halfsine_increase module."""
+
+
+from math import cos, pi, sin
+import unittest
+
 import numpy as np
 
 from nnde.differentialequation.examples.diff3d_halfsine_increase import (
-    D,
-    G, dG_dY, dG_ddelY, dG_ddel2Y,
-    bc, f0, f1, g0, g1, h0, h1, Y0, delbc, del2bc,
+    a, D,
+    G,
+    f0, f1, g0, g1, h0, h1, Y0, bc,
+    df0_dx, df0_dy, df0_dz, df0_dt, df1_dx, df1_dy, df1_dz, df1_dt,
+    dg0_dx, dg0_dy, dg0_dz, dg0_dt, dg1_dx, dg1_dy, dg1_dz, dg1_dt,
+    dh0_dx, dh0_dy, dh0_dz, dh0_dt, dh1_dx, dh1_dy, dh1_dz, dh1_dt,
+    dY0_dx, dY0_dy, dY0_dz, dY0_dt, delbc,
+    d2f0_dx2, d2f0_dy2, d2f0_dz2, d2f0_dt2, d2f1_dx2, d2f1_dy2, d2f1_dz2,
+    d2f1_dt2,
+    d2g0_dx2, d2g0_dy2, d2g0_dz2, d2g0_dt2, d2g1_dx2, d2g1_dy2, d2g1_dz2,
+    d2g1_dt2,
+    d2h0_dx2, d2h0_dy2, d2h0_dz2, d2h0_dt2, d2h1_dx2, d2h1_dy2, d2h1_dz2,
+    d2h1_dt2,
+    d2Y0_dx2, d2Y0_dy2, d2Y0_dz2, d2Y0_dt2, del2bc,
+    dG_ddY_dx, dG_ddY_dy, dG_ddY_dz, dG_ddY_dt, dG_ddelY,
+    dG_dd2Y_dx2, dG_dd2Y_dy2, dG_dd2Y_dz2, dG_dd2Y_dt2, dG_ddel2Y,
     A, delA, del2A,
 )
 
 
+# Grid points for testing.
+n = 11
+xx = np.linspace(0, 1, n)
+yy = np.linspace(0, 1, n)
+zz = np.linspace(0, 1, n)
+tt = np.linspace(0, 1, n)
+
+
+class TestBuilder(unittest.TestCase):
+    """Tests for the diff3d_halfsine_increase module."""
+
+    def test_G(self):
+        """Test the differential equation."""
+        Yt = 2
+        delYt = (3, 4, 5, 6)
+        del2Yt = (5, 6, 7, 8)
+        for x in xx:
+            for y in yy:
+                for z in zz:
+                    for t in tt:
+                        xyzt = (x, y, z, t)
+                        G_ref = (
+                            delYt[3] - D*(del2Yt[0] + del2Yt[1] + del2Yt[2]))
+                        self.assertAlmostEqual(G(xyzt, Yt, delYt, del2Yt),
+                                               G_ref)
+
+    def test_f0(self):
+        """Test boundary condition at (x, y, z, t) = (0, y, z, t)."""
+        f0_ref = 0  # For all inputs.
+        for x in xx:
+            for y in yy:
+                for z in zz:
+                    for t in tt:
+                        xyzt = (x, y, z, t)
+                        self.assertAlmostEqual(f0(xyzt), f0_ref)
+
+    def test_f1(self):
+        """Test boundary condition at (x, y, z, t) = (1, y, z, t)."""
+        for x in xx:
+            for y in yy:
+                for z in zz:
+                    for t in tt:
+                        xyzt = (x, y, z, t)
+                        f1_ref = a*t*sin(pi*y)*sin(pi*z)
[…]
