### 2021-08-23 #3465 @diehlpk
> I made the first review pass at @elwinter's `nnde`. As @hayesall did a great job on creating on-point issues, I went through the functionality check first. Executing this:
>
> ```
> pip install nnde
> git clone https://github.com/elwinter/nnde_demos
> cd nnde_demos; python lagaris01_demo.py
> ```
>
> (`--update` does not work the first time), the demo was successful. However, the demos are located in another repository.
>
> I also ran the `tests` and they all pass.
> I will test it further after [elwinter/nnde#9](https://github.com/elwinter/nnde/issues/9) is resolved.

@taless474 I think #9 is resolved and you can start to review again.
