### 2021-09-06 #3465 @hayesall
Hey @diehlpk I checked off all but "_Functionality_" from my tasklist a couple days ago (but I didn't want to cause notification noise). I still need a little more time on functionality.

Follow-up: This would be a _stronger_ submission with sphinx/mkdocs and https://github.com/elwinter/nnde/issues/13, but nnde currently meets the baseline requirements.
