### 2021-09-06 4a0a7d9
👷 Set test workflow to run on pull requests
Add rules for running Python package tests on pushes to the master
branch and on pull requests that target merging on the master branch.
