+++ .github/workflows/github-run-nnde-tests.yml
@@ -1,5 +1,9 @@
 name: Run nnde regression tests
-on: [push]
+on:
+  push:
+    branches: [ master ]
+  pull_request:
+    branches: [ master ]
 jobs:
   Explore-GitHub-Actions:
     runs-on: ubuntu-latest
