### 2021-09-10 #3465 @diehlpk
@elwinter Can you have a look at  elwinter/nnde#13 and sphinx/mkdocs?
