### 2021-09-13 #3465 @hayesall
@diehlpk Sorry, my previous statement was vague. Those things would be *good* to have (and should probably be done in the future), but I think this is satisfactory *without* them.

Update on functionality review: I've used `nnde` to reproduce some numerical results, some notes are included here https://github.com/elwinter/nnde_demos/pull/4
