### 2021-09-14 elwinter/nnde#22 @taless474
Installation instruction: --update
To install and use the software, the [documentation](https://github.com/elwinter/nnde) begins with `pip install nnde --update`. This option does not work when a user install the software for the first time.
```console
PS C:\Users\Taless\Desktop\nnde> pip install nnde --update

Usage:
  pip install [options] <requirement specifier> [package-index-options] ...
  pip install [options] -r <requirements file> [package-index-options] ...
  pip install [options] [-e] <vcs project url> ...
  pip install [options] [-e] <local project path> ...
  pip install [options] <archive url/path> ...

no such option: --update
```
