### 2021-09-14 #3465 @taless474
@elwinter, thank you for addressing the issues. I started checking the installation and created [#22](https://github.com/elwinter/nnde/issues/22). Would you please take a look at it?

@diehlpk, do we need the demo to be in the same repository? I am not sure if this is an issue and thank you in advance for your response.
