### 2021-09-19 d0b18ae
Added several new example problems from Lagaris.
