+++ nnde/differentialequation/examples/lagaris_05.py
@@ -1,0 +1,644 @@
+"""
+This module implements problem 5 in Lagaris et al (1998) (2nd order 2-D
+PDE BVP).
+
+Note that an upper-case 'Y' is used to represent the Greek psi from
+the original equation.
+
+The equation is defined on the domain [[0, 1], [0, 1]].
+
+The analytical form of the equation is:
+
+    G(x, y, Y, dY/dx, dY/dy, d2Y/dx2, d2Y/dy2) =
+    d2Y_dx2 + d2Y_dy2 - exp(-x) (x - 2 + y**3 + 6*y) = 0
+
+with initial condition:
+
+    Y(0, y) = y**3
+    Y(1, y) = (1 + y**3)*exp(-1)
+    Y(x, 0) = x*exp(-x)
+    Y(x, 1) = exp(-x)*(x + 1)
+
+This equation has the analytical solution for the supplied initial
+conditions:
+
+Ya(x, y) = exp(-x) (x + y**3)
+
+Reference:
+
+Isaac Elias Lagaris, Aristidis Likas, and Dimitrios I. Fotiadis,
+"Artificial Neural Networks for Solving Ordinary and Partial Differential
+Equations", *IEEE Transactions on Neural Networks* **9**(5), pp. 987-999,
+1998
+"""
+
+
+from math import exp
+
+
+def G(xy, Y, delY, del2Y):
+    """Compute the differential equation in standard form.
+
+    Compute the value of the differential equation in standard form. For a
+    perfect solution, the value should be 0.
+
+    Parameters
+    ----------
+    xy : array-like of 2 float
+        Values for x and y, in that order.
+    Y : float
+        Current solution value.
+    delY : array-like of 2 float, not used
+        Values for dY/dx and dY/dy, in that order.
+    del2Y : array-like of 2 float
+        Values for d2Y/dx2 and d2Y/dy2, in that order.
+
+    Returns
+    -------
+    result : float
+        Value of the differential equation.
+    """
+    (x, y) = xy
+    (d2Y_dx2, d2Y_dy2) = del2Y
+    _G = d2Y_dx2 + d2Y_dy2 - exp(-x)*(x - 2 + y**3 + 6*y)
+    return _G
+
+
+# def f0(xt):
+#     """Boundary condition at (x, t) = (0, t).
+
+#     Compute the value of the solution at (x, t) = (0, t).
+
+#     Parameters
+#     ----------
+#     xt : array-like of 2 float
+#         Values for x and t, in that order.
+
+#     Returns
+#     -------
+#     result : float
+#         Value of the solution at (x, t) = (0, t).
+#     """
+#     return C
+
+
+# def f1(xt):
+#     """Boundary condition at (x, t) = (1, t).
+
+#     Compute the value of the solution at (x, t) = (1, t).
+
+#     Parameters
+#     ----------
+#     xt : array-like of 2 float
+#         Values for x and t, in that order.
+
+#     Returns
+#     -------
+#     result : float
+#         Value of the solution at (x, t) = (1, t).
+#     """
+#     return C
+
+
+# def Y0(xt):
+#     """Boundary condition at (x, t) = (x, 0).
+
+#     Compute the value of the solution at (x, t) = (x, 0).
+
+#     Parameters
+#     ----------
+#     xt : array-like of 2 float
+#         Values for x and t, in that order.
+
+#     Returns
+#     -------
+#     result : float
+#         Value of the solution at (x, t) = (x, 0).
+#     """
+#     return C
+
+
+# # Gather the boundary condition functions in a single array.
+# bc = [[f0, f1], [Y0, None]]
+
+
+# def df0_dx(xt):
+#     """1st derivative of BC wrt x at (x, t) = (0, t).
+
+#     Compute the 1st derivative of BC wrt x at (x, t) = (0, t).
+
[…]

[…]
