### 2021-10-04 elwinter/nnde#22 @diehlpk
@elwinter Could you please have a look?
