### 2021-10-04 #3465 @taless474
@diehlpk, thank you for your comment. We are waiting for @elwinter to address [#22](https://github.com/elwinter/nnde/issues/22), so I can go forward with the next round of review.
