### 2021-10-15 #3465 @elwinter
Hmm. I thought I had fixed and closed all but one issue a couple of weeks ago, and was waiting for a response on my test coverage question.  Did i miss something? Please let me know and i will take care of it.

Thanks
Eric

Get Outlook for Android<https://aka.ms/ghei36>
