### 2021-10-15 elwinter/nnde#13 @elwinter
Ok that is what I wanted to know. I will be working on that this weekend.

Thanks
Eric

Get Outlook for Android<https://aka.ms/ghei36>
