### 2021-10-23 c6ab049
Added more demo cases.
