+++ nnde/differentialequation/examples/pde_01.py
@@ -1,39 +1,40 @@
 """
-This module implements problem 5 in Lagaris et al (1998) (2nd order 2-D
-PDE BVP).
+This module implements a simple 2-D 1st-order PDE IVP.
 
 Note that an upper-case 'Y' is used to represent the Greek psi from
 the original equation.
 
-The equation is defined on the domain [[0, 1], [0, 1]].
+The equation is defined on the domain [[0, 1], [0, 1]]:
+
+    dY_dx + dY_dy = y**2 + 2*x*y
 
 The analytical form of the equation is:
 
-    G(x, y, Y, dY/dx, dY/dy, d2Y/dx2, d2Y/dy2) =
-    d2Y_dx2 + d2Y_dy2 - exp(-x)*(x - 2 + y**3 + 6*y) = 0
+    G(x, y, Y, dY/dx, dY/dy) = 0 =
+    dY_dx + dY_dy - y**2 - 2*x*y
 
-with initial condition:
+with initial conditions:
 
-    Y(0, y) = y**3
-    Y(1, y) = (1 + y**3)*exp(-1)
-    Y(x, 0) = x*exp(-x)
-    Y(x, 1) = exp(-x)*(x + 1)
+    Y(0, y) = f0(x, y) = 0
+    Y(x, 0) = g0(x, y) = 0
 
 This equation has the analytical solution for the supplied initial
 conditions:
 
-Ya(x, y) = exp(-x)*(x + y**3)
+Ya(x, y) = x*y**2
+dYa_dx(x, y) = y**2
+dYa_dy(x, y) = 2*x*y
 
-Reference:
+The boundary condition function is:
 
-Isaac Elias Lagaris, Aristidis Likas, and Dimitrios I. Fotiadis,
-"Artificial Neural Networks for Solving Ordinary and Partial Differential
-Equations", *IEEE Transactions on Neural Networks* **9**(5), pp. 987-999,
-1998
+A(x, y) = (1 - x)*f0(x, y)
+          + (1 - y)*(g0(x, y) - ((1 - x)*g0(x, 0) + x*g0(x, 1)))
+        = 0
+
+The trial solution is:
+
+Yt(x, y) = x*y*N(x, y, p)
 """
-
-
-from math import exp
 
 
 def G(xy, Y, delY, del2Y):
@@ -48,7 +49,7 @@ def G(xy, Y, delY, del2Y):
         Values for x and y, in that order.
     Y : float
         Current solution value.
-    delY : array-like of 2 float, not used
+    delY : array-like of 2 float
         Values for dY/dx and dY/dy, in that order.
     del2Y : array-like of 2 float
         Values for d2Y/dx2 and d2Y/dy2, in that order.
@@ -59,8 +60,9 @@ def G(xy, Y, delY, del2Y):
         Value of the differential equation.
     """
     (x, y) = xy
+    (dY_dx, dY_dy) = delY
     (d2Y_dx2, d2Y_dy2) = del2Y
-    _G = d2Y_dx2 + d2Y_dy2 - exp(-x)*(x - 2 + y**3 + 6*y)
+    _G = dY_dx + dY_dy - y**2 - 2*x*y
     return _G
 
 
@@ -80,7 +82,7 @@ def f0(xy):
         Value of the solution at (x, y) = (0, y).
     """
     (x, y) = xy
-    return y**3
+    return 0
 
 
 def f1(xy):
@@ -99,7 +101,7 @@ def f1(xy):
         Value of the solution at (x, y) = (1, y).
     """
     (x, y) = xy
-    return (1 + y**3)*exp(-1)
+    return y**2
 
 
 def g0(xy):
@@ -118,7 +120,7 @@ def g0(xy):
[…]
