### 2021-10-23 be6ac08
Added second version of training data function.
