### 2021-11-06 1161e92
Removed unused code.
