+++ nnde/differentialequation/examples/pde_02.py
@@ -1,677 +1,0 @@
-"""
-This module implements a simple 2-D 2nd-order PDE BVP.
-
-Note that an upper-case 'Y' is used to represent the Greek psi from
-the original equation.
-
-The equation is defined on the domain [[0, 1], [0, 1]]:
-
-    d2Y_dx2 + d2Y_dy2 = 2*x
-
-The analytical form of the equation is:
-
-    G(x, y, Y, dY/dx, dY/dy) = 0 =
-    d2Y_dx2 + d2Y_dy2 - 2*x
-
-with boundary conditions:
-
-    Y(0, y) = f0(x, y) = 0
-    Y(1, y) = f1(x, y) = y**2
-    Y(x, 0) = g0(x, y) = 0
-    Y(x, 0) = g1(x, y) = x
-
-This equation has the analytical solution for the supplied boundary
-conditions:
-
-Ya(x, y) = x*y**2
-dYa_dx(x, y) = y**2
-dYa_dy(x, y) = 2*x*y
-d2Ya_dx2(x, y) = 0
-d2Ya_dy2(x, y) = 2*x
-
-The boundary condition function is:
-
-A(x, y) = (1 - x)*f0(x, y) + x * f1(x, y)
-          + (1 - y)*(g0(x, y) - ((1 - x)*g0(0, 0) + x*g0(1, 0)))
-          + y*(g1(x, y) - ((1 - x)*g1(0, y) + x*g1(1, y)))
-        = 0 + x * y**2
-          + 0
-          + y*(x - ((1 - x)*0 + x*1))
-        = x + y**2
-
-The trial solution is:
-
-Yt(x, y) = A(x, y) + x*(1 - x)*y*(1 - y)*N(x, y, p)
-         = x + y**2 + x*(1 - x)*y*(1 - y)*N(x, y, p)
-"""
-
-
-def G(xy, Y, delY, del2Y):
-    """Compute the differential equation in standard form.
-
-    Compute the value of the differential equation in standard form. For a
-    perfect solution, the value should be 0.
-
-    Parameters
-    ----------
-    xy : array-like of 2 float
-        Values for x and y, in that order.
-    Y : float
-        Current solution value.
-    delY : array-like of 2 float
-        Values for dY/dx and dY/dy, in that order.
-    del2Y : array-like of 2 float
-        Values for d2Y/dx2 and d2Y/dy2, in that order.
-
-    Returns
-    -------
-    result : float
-        Value of the differential equation.
-    """
-    (x, y) = xy
-    (dY_dx, dY_dy) = delY
-    (d2Y_dx2, d2Y_dy2) = del2Y
-    _G = d2Y_dx2 + d2Y_dy2 - 2*x
-    return _G
-
-
-def f0(xy):
-    """Boundary condition at (x, y) = (0, y).
-
-    Compute the value of the solution at (x, y) = (0, y).
-
-    Parameters
-    ----------
-    xy : array-like of 2 float
-        Values for x and y, in that order.
-
-    Returns
-    -------
-    result : float
-        Value of the solution at (x, y) = (0, y).
-    """
-    (x, y) = xy
-    return 0
-
-
-def f1(xy):
-    """Boundary condition at (x, y) = (1, y).
-
-    Compute the value of the solution at (x, y) = (1, y).
-
-    Parameters
-    ----------
-    xy : array-like of 2 float
-        Values for x and y, in that order.
-
-    Returns
-    -------
-    result : float
-        Value of the solution at (x, y) = (1, y).
-    """
-    (x, y) = xy
-    return y**2
-
-
-def g0(xy):
-    """Boundary condition at (x, y) = (x, 0).
-
-    Compute the value of the solution at (x, y) = (x, 0).
-
-    Parameters
[…]
