### 2021-11-06 #3465 @elwinter
Hi Patrick. I apologize for the delayed response. As shown below, test
coverage (the only remaining issue) is now at 76%. Is this sufficient to
continue?

Thanks
Eric

Coverage report: 76%[image: Show keyboard shortcuts]
Modulestatementsmissingexcludedcoverage
Total 6341 1508 0 76%
nnde/__init__.py <nnde___init___py.html> 0 0 0 100%
nnde/differentialequation/__init__.py
<nnde_differentialequation___init___py.html> 0 0 0 100%
nnde/differentialequation/differentialequation.py
<nnde_differentialequation_differentialequation_py.html> 6 0 0 100%
nnde/differentialequation/examples/__init__.py
<nnde_differentialequation_examples___init___py.html> 0 0 0 100%
nnde/differentialequation/examples/diff1d_half.py
<nnde_differentialequation_examples_diff1d_half_py.html> 69 0 0 100%
nnde/differentialequation/examples/diff1d_halfsine.py
<nnde_differentialequation_examples_diff1d_halfsine_py.html> 87 0 0 100%
nnde/differentialequation/examples/diff1d_halfsine_increase.py
<nnde_differentialequation_examples_diff1d_halfsine_increase_py.html> 106 0
0 100%
nnde/differentialequation/examples/diff1d_one.py
<nnde_differentialequation_examples_diff1d_one_py.html> 70 0 0 100%
nnde/differentialequation/examples/diff1d_zero.py
<nnde_differentialequation_examples_diff1d_zero_py.html> 70 0 0 100%
nnde/differentialequation/examples/diff2d_halfsine.py
<nnde_differentialequation_examples_diff2d_halfsine_py.html> 142 7 0 95%
nnde/differentialequation/examples/diff2d_halfsine_increase.py
<nnde_differentialequation_examples_diff2d_halfsine_increase_py.html> 122 1
0 99%
nnde/differentialequation/examples/diff3d_halfsine.py
<nnde_differentialequation_examples_diff3d_halfsine_py.html> 213 118 0 45%
nnde/differentialequation/examples/diff3d_halfsine_increase.py
<nnde_differentialequation_examples_diff3d_halfsine_increase_py.html> 188 11
0 94%
nnde/differentialequation/examples/example_pde1ivp_01.py
<nnde_differentialequation_examples_example_pde1ivp_01_py.html> 53 28 0 47%
nnde/differentialequation/examples/lagaris_01.py
<nnde_differentialequation_examples_lagaris_01_py.html> 13 5 0 62%
nnde/differentialequation/examples/lagaris_02.py
<nnde_differentialequation_examples_lagaris_02_py.html> 13 5 0 62%
nnde/differentialequation/examples/tests/__init__.py
<nnde_differentialequation_examples_tests___init___py.html> 0 0 0 100%
nnde/differentialequation/examples/tests/test_diff1d_half.py
<nnde_differentialequation_examples_tests_test_diff1d_half_py.html> 263 1 0
99%
nnde/differentialequation/examples/tests/test_diff1d_halfsine.py
<nnde_differentialequation_examples_tests_test_diff1d_halfsine_py.html> 262
1 0 99%
nnde/differentialequation/examples/tests/test_diff1d_halfsine_increase.py
<nnde_differentialequation_examples_tests_test_diff1d_halfsine_increase_py.html>
272 1 0 99%
nnde/differentialequation/examples/tests/test_diff1d_one.py
<nnde_differentialequation_examples_tests_test_diff1d_one_py.html> 263 1 0
99%
nnde/differentialequation/examples/tests/test_diff1d_zero.py
<nnde_differentialequation_examples_tests_test_diff1d_zero_py.html> 263 1 0
99%
nnde/differentialequation/examples/tests/test_diff2d_halfsine.py
<nnde_differentialequation_examples_tests_test_diff2d_halfsine_py.html> 493
11 0 98%
nnde/differentialequation/examples/tests/test_diff2d_halfsine_increase.py
<nnde_differentialequation_examples_tests_test_diff2d_halfsine_increase_py.html>
428 1 0 99%
nnde/differentialequation/examples/tests/test_diff3d_halfsine.py
<nnde_differentialequation_examples_tests_test_diff3d_halfsine_py.html> 92
89 0 3%
nnde/differentialequation/examples/tests/test_diff3d_halfsine_increase.py
<nnde_differentialequation_examples_tests_test_diff3d_halfsine_increase_py.html>
710 1 0 99%
nnde/differentialequation/examples/tests/test_example_pde1ivp_01.py
<nnde_differentialequation_examples_tests_test_example_pde1ivp_01_py.html>
35 32 0 9%
nnde/differentialequation/examples/tests/test_lagaris_01.py
<nnde_differentialequation_examples_tests_test_lagaris_01_py.html> 8 5 0 38%
nnde/differentialequation/examples/tests/test_lagaris_02.py
<nnde_differentialequation_examples_tests_test_lagaris_02_py.html> 8 5 0 38%
nnde/differentialequation/ode/__init__.py
<nnde_differentialequation_ode___init___py.html> 0 0 0 100%
nnde/differentialequation/ode/ode.py
<nnde_differentialequation_ode_ode_py.html> 3 0 0 100%
nnde/differentialequation/ode/ode1.py
<nnde_differentialequation_ode_ode1_py.html> 5 0 0 100%
nnde/differentialequation/ode/ode1ivp.py
<nnde_differentialequation_ode_ode1ivp_py.html> 27 14 0 48%
nnde/differentialequation/ode/tests/__init__.py
<nnde_differentialequation_ode_tests___init___py.html> 0 0 0 100%
nnde/differentialequation/ode/tests/test_ODE.py
<nnde_differentialequation_ode_tests_test_ODE_py.html> 9 1 0 89%
nnde/differentialequation/ode/tests/test_ODE1.py
<nnde_differentialequation_ode_tests_test_ODE1_py.html> 12 1 0 92%
nnde/differentialequation/ode/tests/test_ODE1IVP.py
<nnde_differentialequation_ode_tests_test_ODE1IVP_py.html> 7 1 0 86%
nnde/differentialequation/pde/__init__.py
<nnde_differentialequation_pde___init___py.html> 0 0 0 100%
nnde/differentialequation/pde/pde.py
<nnde_differentialequation_pde_pde_py.html> 3 0 0 100%
nnde/differentialequation/pde/pde1.py
<nnde_differentialequation_pde_pde1_py.html> 5 0 0 100%
nnde/differentialequation/pde/pde1ivp.py
<nnde_differentialequation_pde_pde1ivp_py.html> 29 16 0 45%
nnde/differentialequation/pde/pde2.py
<nnde_differentialequation_pde_pde2_py.html> 5 0 0 100%
nnde/differentialequation/pde/pde2diff.py
<nnde_differentialequation_pde_pde2diff_py.html> 47 28 0 40%
nnde/differentialequation/pde/tests/__init__.py
<nnde_differentialequation_pde_tests___init___py.html> 0 0 0 100%
nnde/differentialequation/pde/tests/test_PDE.py
<nnde_differentialequation_pde_tests_test_PDE_py.html> 9 1 0 89%
nnde/differentialequation/pde/tests/test_PDE1.py
<nnde_differentialequation_pde_tests_test_PDE1_py.html> 12 1 0 92%
nnde/differentialequation/pde/tests/test_PDE1IVP.py
<nnde_differentialequation_pde_tests_test_PDE1IVP_py.html> 8 1 0 88%
nnde/differentialequation/pde/tests/test_PDE2.py
<nnde_differentialequation_pde_tests_test_PDE2_py.html> 12 1 0 92%
nnde/differentialequation/pde/tests/test_PDE2DIFF.py
<nnde_differentialequation_pde_tests_test_PDE2DIFF_py.html> 7 1 0 86%
nnde/differentialequation/tests/__init__.py
<nnde_differentialequation_tests___init___py.html> 0 0 0 100%
nnde/differentialequation/tests/test_DifferentialEquation.py
<nnde_differentialequation_tests_test_DifferentialEquation_py.html> 12 1 0
92%
nnde/exceptions/__init__.py <nnde_exceptions___init___py.html> 0 0 0 100%
nnde/exceptions/nndeexception.py <nnde_exceptions_nndeexception_py.html> 2 0
0 100%
nnde/exceptions/tests/__init__.py <nnde_exceptions_tests___init___py.html> 0
0 0 100%
nnde/exceptions/tests/test_NNDEException.py
<nnde_exceptions_tests_test_NNDEException_py.html> 7 1 0 86%
nnde/math/__init__.py <nnde_math___init___py.html> 0 0 0 100%
nnde/math/kdelta.py <nnde_math_kdelta_py.html> 2 0 0 100%
nnde/math/sigma.py <nnde_math_sigma_py.html> 22 0 0 100%
nnde/math/tests/__init__.py <nnde_math_tests___init___py.html> 0 0 0 100%
nnde/math/tests/test_kdelta.py <nnde_math_tests_test_kdelta_py.html> 11 1 0
91%
nnde/math/tests/test_sigma.py <nnde_math_tests_test_sigma_py.html> 37 1 0
97%
nnde/math/tests/test_trainingdata.py
<nnde_math_tests_test_trainingdata_py.html> 27 1 0 96%
nnde/math/trainingdata.py <nnde_math_trainingdata_py.html> 41 16 0 61%
nnde/neuralnetwork/__init__.py <nnde_neuralnetwork___init___py.html> 0 0 0
100%
nnde/neuralnetwork/neuralnetwork.py
<nnde_neuralnetwork_neuralnetwork_py.html> 8 0 0 100%
nnde/neuralnetwork/nnode1ivp.py <nnde_neuralnetwork_nnode1ivp_py.html> 250
214 0 14%
nnde/neuralnetwork/nnpde1ivp.py <nnde_neuralnetwork_nnpde1ivp_py.html> 474
425 0 10%
nnde/neuralnetwork/nnpde2diff.py <nnde_neuralnetwork_nnpde2diff_py.html> 490
448 0 9%
nnde/neuralnetwork/slffnn.py <nnde_neuralnetwork_slffnn_py.html> 3 0 0 100%
nnde/neuralnetwork/tests/__init__.py
<nnde_neuralnetwork_tests___init___py.html> 0 0 0 100%
nnde/neuralnetwork/tests/test_NNODE1IVP.py
<nnde_neuralnetwork_tests_test_NNODE1IVP_py.html> 29 1 0 97%
nnde/neuralnetwork/tests/test_NNPDE1IVP.py
<nnde_neuralnetwork_tests_test_NNPDE1IVP_py.html> 7 1 0 86%
nnde/neuralnetwork/tests/test_NeuralNetwork.py
<nnde_neuralnetwork_tests_test_NeuralNetwork_py.html> 15 1 0 93%
nnde/neuralnetwork/tests/test_PDE2DIFF.py
<nnde_neuralnetwork_tests_test_PDE2DIFF_py.html> 27 1 0 96%
nnde/neuralnetwork/tests/test_SLFFNN.py
<nnde_neuralnetwork_tests_test_SLFFNN_py.html> 9 1 0 89%
nnde/trialfunction/__init__.py <nnde_trialfunction___init___py.html> 0 0 0
100%
nnde/trialfunction/diff1dtrialfunction.py
<nnde_trialfunction_diff1dtrialfunction_py.html> 73 0 0 100%
nnde/trialfunction/diff2dtrialfunction.py
<nnde_trialfunction_diff2dtrialfunction_py.html> 79 0 0 100%
nnde/trialfunction/diff3dtrialfunction.py
<nnde_trialfunction_diff3dtrialfunction_py.html> 85 0 0 100%
nnde/trialfunction/tests/__init__.py
<nnde_trialfunction_tests___init___py.html> 0 0 0 100%
nnde/trialfunction/tests/test_Diff1DTrialFunction.py
<nnde_trialfunction_tests_test_Diff1DTrialFunction_py.html> 57 1 0 98%
nnde/trialfunction/tests/test_Diff2DTrialFunction.py
<nnde_trialfunction_tests_test_Diff2DTrialFunction_py.html> 57 1 0 98%
nnde/trialfunction/tests/test_Diff3DTrialFunction.py
<nnde_trialfunction_tests_test_Diff3DTrialFunction_py.html> 57 1 0 98%
nnde/trialfunction/tests/test_TrialFunction.py
<nnde_trialfunction_tests_test_TrialFunction_py.html> 8 1 0 88%
nnde/trialfunction/trialfunction.py
<nnde_trialfunction_trialfunction_py.html> 3 1 0 67%

coverage.py v5.5 <https://coverage.readthedocs.io>, created at 2021-11-06
10:13 -0400

On Mon, Nov 1, 2021 at 11:17 AM Patrick Diehl ***@***.***>
wrote:

> @elwinter <https://github.com/elwinter> Please address the editorial
> comments and we are good to accept the paper.
>
