### 2021-11-06 #3465 @hayesall
I just tested the master branch and got 72% coverage.

<summary>Terminal Recap</summary>

```bash
git clone git@github.com:elwinter/nnde.git
cd nnde
pip install pytest pytest-cov
pip install -e .
pytest --cov=nnde .
# ---
# TOTAL                                                                        6550   1857    72%
```

This seems fine (I tend to think coverage numbers are more a rule-of-thumb than an absolute threshold).

If you want to report coverage stats on push/pull requests, you can add something like this to the `.github/workflows/github-run-nnde-tests.yml`:

```yml
    - name: Test with pytest
      run:  pytest --cov=nnde .
    - name: Upload coverage to Codecov
      uses: codecov/codecov-action@v1
```
