### 2021-11-07 elwinter/nnde#13 @diehlpk
@elwinter I think we can close the issue since you fixed the coverage issue.
