### 2021-11-07 #3465 @diehlpk
> @elwinter I briefly read the paper and have the following remarks, so far
>
>     * [ ]  Please add the city, state, and country to the affiliation
>
>     * [ ]  Provide some reference to PyTorch and TensorFlow 2. In addition, I think the related work section is really sparse and some mention of other packages and how they differ would be beneficial for the reader.
>
>     * [ ]  You mention that FEM and FDM can be difficult to parallelize. Can you provide some reference for that claim?

@winter I think these are the last open items for the paper.
