### 2021-11-07 #3465 @elwinter
Ok are there issues open for them, or are they to be handled outside the issue flow?
Thanks Eric

Get Outlook for Android<https://aka.ms/ghei36>
