### 2021-11-08 #3465 @elwinter
OK thanks, Just wanted to make sure I was not missing anything.

Eric

On Sun, Nov 7, 2021 at 8:37 PM Patrick Diehl ***@***.***>
wrote:

> Ok are there issues open for them, or are they to be handled outside the
> issue flow? Thanks Eric Get Outlook for Androidhttps://aka.ms/ghei36
> … <#m_990835679522565801_>
> ________________________________ From: Patrick Diehl *@*.*> Sent: Sunday,
> November 7, 2021 3:19:31 PM To: openjournals/joss-reviews @.*> Cc: Eric
> Winter *@*.*>; Mention @.*> Subject: Re: [openjournals/joss-reviews]
> [REVIEW]: nnde: A Python package for solving differential equations using
> neural networks (#3465
> <https://github.com/openjournals/joss-reviews/issues/3465>) @elwinter
> <https://github.com/elwinter>https://github.com/elwinter I briefly read
> the paper and have the following remarks, so far * [ ] Please add the city,
> state, and country to the affiliation * [ ] Provide some reference to
> PyTorch and TensorFlow 2. In addition, I think the related work section is
> really sparse and some mention of other packages and how they differ would
> be beneficial for the reader. * [ ] You mention that FEM and FDM can be
> difficult to parallelize. Can you provide some reference for that claim?
> @Winter <https://github.com/Winter>https://github.com/Winter I think
> these are the last open items for the paper. — You are receiving this
> because you were mentioned. Reply to this email directly, view it on GitHub<#3465
> (comment)
> <https://github.com/openjournals/joss-reviews/issues/3465#issuecomment-962674280>>,
> or unsubscribe
> https://github.com/notifications/unsubscribe-auth/ABSFWC4Y6MD7IZFBM2K3NZTUK3NNHANCNFSM5AC2IOEQ.
> Triage notifications on the go with GitHub Mobile for iOS
> https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&mt=8&pt=524675
> or Android
> https://play.google.com/store/apps/details?id=com.github.android&referrer=utm_campaign%3Dnotification-email%26utm_medium%3Demail%26utm_source%3Dgithub
> .
>
> I just document the three issues above in this thread.
>
