+++ paper.md
@@ -12,7 +12,7 @@ authors:
     orcid: 0000-0002-9521-5228
     affiliation: 1
 affiliations:
-  - name: Department of Physics and Astronomy, George Mason University
+  - name: Department of Physics and Astronomy, George Mason University, Fairfax, Virginia, USA
     index: 1
 date: 14 August 2021
 bibliography: paper.bib
@@ -26,9 +26,9 @@ Neural networks have been shown to have the ability to solve differential equati
 
 The `nnde` package provides a pure-Python implementation of one of the earliest approaches to using neural networks to solve differential equations - the trial function method [@Lagaris:1998]. The `nnde` package was initially developed as a vehicle for understanding the internal workings of feedforward neural networks, without the constraints imposed by an existing neural network framework. It has since been enhanced to provide the capability to solve differential equations of scientific interest, such as the diffusion equation described here. The ultimate goal of the package is to provide the capability to solve systems of coupled partial differential equations, such as the equations of magnetohydrodynamics.
 
-Development of the `nnde` package began before the widespread adoption of modern neural network software frameworks, such as TensorFlow 2 and PyTorch. These frameworks provide autodifferentiation capability - by recording the sequence of mathematical operations performed in the forward pass through the network, the framework can automatically compute the gradients of the loss function with respect to each of the network parameters, as well as the network inputs. The latter capability is central to solving differential equations. The `nnde` package uses precomputed derivative functions for the components of the differential equation and the trial solution.
+Development of the `nnde` package began before the widespread adoption of modern neural network software frameworks, such as TensorFlow 2 (`https://tensorflow.org`) and PyTorch (`https://pytorch.org`). These frameworks provide autodifferentiation capability - by recording the sequence of mathematical operations performed in the forward pass through the network, the framework can automatically compute the gradients of the loss function with respect to each of the network parameters, as well as the network inputs. The latter capability is central to solving differential equations. The `nnde` package uses precomputed derivative functions for the components of the differential equation and the trial solution.
 
-The most commonly used methods for solving differential equations are the Finite Element Method (FEM) and Finite Difference Method (FDM). However, these methods can be difficult to parallelize and may have large storage requirements for model outputs. The neural network method is straightforward to parallelize due to the independent characteristics of the computational nodes in each network layer. Additionally, the trained network solution is more compact than an FDM or FEM solution because storage of only the network weights and biases is required. The neural network solution is mesh-free and does not require interpolation to retrieve the solution at a non-grid point, as is the case with FDM or FEM. Once the network is trained, computing a solution at any spatial or temporal scale requires only a series of matrix multiplications, one per network layer. The trained solution is a sum of arbitrary differentiable basis functions, and therefore the trained solution is also differentiable, which is particularly useful when computing derived quantities such as gradients and fluxes.
[…]
