+++ paper.bib
@@ -26,6 +26,18 @@
   journal = {{IEEE} Transactions on Neural Networks}
 }
 
+@article{Raissi:2019,
+  doi = {10.1016/j.jcp.2018.10.045},
+  year  = {2019},
+  publisher = {Elsevier},
+  volume = {378},
+  number = {},
+  pages = {686--707},
+  author = {M. Raissi and P. Perdikaris and G.E. Karniadakis},
+  title = {Physics-informed neural networks: A deep learning framework for solving forward and inverse problems involving nonlinear partial differential equations},
+  journal = {Journal of Computational Physics}
+}
+
 @book{Yadav:2015,
   doi = {10.1007/978-94-017-9816-7},
   year = {2015},
