### 2021-11-12 #3465 @elwinter
Hi Patrick. I have added some text, references, and links to the paper.md
that I think address these concerns. I still did not spend a lot of text on
other software, due to the space limitation. If you feel this still needs
to be changed, please let me know.

I verified that the paper.md & paper.bib render properly in whedon. Please
let me know if you or the other reviewers have any questions.

Thanks
Eric Winter

On Sun, Nov 7, 2021 at 3:19 PM Patrick Diehl ***@***.***>
wrote:

> @elwinter <https://github.com/elwinter> I briefly read the paper and have
> the following remarks, so far
>
> * [ ]  Please add the city, state, and country to the affiliation
>
> * [ ]  Provide some reference to PyTorch and TensorFlow 2. In addition, I think the related work section is really sparse and some mention of other packages and how they differ would be beneficial for the reader.
>
> * [ ]  You mention that FEM and FDM can be difficult to parallelize. Can you provide some reference for that claim?
>
> @Winter <https://github.com/Winter> I think these are the last open items
> for the paper.
>
