### 2021-11-15 #3465 @diehlpk
> Hi Patrick. I have added some text, references, and links to the paper.md that I think address these concerns. I still did not spend a lot of text on other software, due to the space limitation. If you feel this still needs to be changed, please let me know. I verified that the paper.md & paper.bib render properly in whedon. Please let me know if you or the other reviewers have any questions. Thanks Eric Winter
> […](#)
> On Sun, Nov 7, 2021 at 3:19 PM Patrick Diehl ***@***.***> wrote: @elwinter <https://github.com/elwinter> I briefly read the paper and have the following remarks, so far * [ ] Please add the city, state, and country to the affiliation * [ ] Provide some reference to PyTorch and TensorFlow 2. In addition, I think the related work section is really sparse and some mention of other packages and how they differ would be beneficial for the reader. * [ ] You mention that FEM and FDM can be difficult to parallelize. Can you provide some reference for that claim? @Winter <https://github.com/Winter> I think these are the last open items for the paper. — You are receiving this because you were mentioned. Reply to this email directly, view it on GitHub <[#3465 (comment)](https://github.com/openjournals/joss-reviews/issues/3465#issuecomment-962674280)>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/ABSFWC4Y6MD7IZFBM2K3NZTUK3NNHANCNFSM5AC2IOEQ> . Triage notifications on the go with GitHub Mobile for iOS <https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&mt=8&pt=524675> or Android <https://play.google.com/store/apps/details?id=com.github.android&referrer=utm_campaign%3Dnotification-email%26utm_medium%3Demail%26utm_source%3Dgithub>.

@elwinter I am fine with all of your other changes. But I still think have some more details on related work will be helpful to the reader. Might you add some list of similar packages and provide some references?

It reads strange that nnde, TensorFlow, and PyTorch are the only packages.

Or at least make some statement that many more packages are available, but TensorFlow and PyTorch are the most simialr ones?
