### 2021-11-28 #3465 @elwinter
Hi Patrick. I thought I had already addressed all of the issues a couple
of weeks ago, but now I see a message from you on 15 November. My apologies
- I did not see that message originally. I will plan to make further
modifications early this week. I too would like to put this one "to
bed".... :^)

Take care
Eric

On Sun, Nov 28, 2021 at 2:38 PM Patrick Diehl ***@***.***>
wrote:

> @elwinter <https://github.com/elwinter> Do you think we can finish the
> paper by the end of next week? I would like to finalize the paper and take
> a new paper.
>
