+++ nnde/differentialequation/examples/lagaris_06.py
@@ -1,0 +1,663 @@
+"""
+This module implements problem 6 in Lagaris et al (1998) (2nd order 2-D
+PDE BVP).
+
+Note that an upper-case 'Y' is used to represent the Greek psi from
+the original equation.
+
+The equation is defined on the domain [[0, 1], [0, 1]].
+
+The analytical form of the equation is:
+
+    G(x, y, Y, dY/dx, dY/dy, d2Y/dx2, d2Y/dy2) =
+    d2Y_dx2 + d2Y_dy2 - (2 - pi***2*y**2)*sin(pi*x) = 0
+
+with boundary conditions:
+
+    Y(0, y) = 0
+    Y(1, y) = 0
+    Y(x, 0) = 0
+    dY/dy(x, 1) = 2*sin(pi*x)
+
+This equation has the analytical solution for the supplied initial
+conditions:
+
+Ya(x, y) = y**2*sin(pi*x)
+
+Reference:
+
+Isaac Elias Lagaris, Aristidis Likas, and Dimitrios I. Fotiadis,
+"Artificial Neural Networks for Solving Ordinary and Partial Differential
+Equations", *IEEE Transactions on Neural Networks* **9**(5), pp. 987-999,
+1998
+"""
+
+
+from math import cos, pi, sin
+
+
+def G(xy, Y, delY, del2Y):
+    """Compute the differential equation in standard form.
+
+    Compute the value of the differential equation in standard form. For a
+    perfect solution, the value should be 0.
+
+    Parameters
+    ----------
+    xy : array-like of 2 float
+        Values for x and y, in that order.
+    Y : float
+        Current solution value.
+    delY : array-like of 2 float, not used
+        Values for dY/dx and dY/dy, in that order.
+    del2Y : array-like of 2 float
+        Values for d2Y/dx2 and d2Y/dy2, in that order.
+
+    Returns
+    -------
+    result : float
+        Value of the differential equation.
+    """
+    (x, y) = xy
+    (d2Y_dx2, d2Y_dy2) = del2Y
+    _G = d2Y_dx2 + d2Y_dy2 - (2 - pi**2*y**2)*sin(pi*x)
+    return _G
+
+
+def f0(xy):
+    """Boundary condition at (x, y) = (0, y).
+
+    Compute the value of the solution at (x, y) = (0, y).
+
+    Parameters
+    ----------
+    xy : array-like of 2 float
+        Values for x and y, in that order.
+
+    Returns
+    -------
+    result : float
+        Value of the solution at (x, y) = (0, y).
+    """
+    return 0
+
+
+def f1(xy):
+    """Boundary condition at (x, y) = (1, y).
+
+    Compute the value of the solution at (x, y) = (1, y).
+
+    Parameters
+    ----------
+    xy : array-like of 2 float
+        Values for x and y, in that order.
+
+    Returns
+    -------
+    result : float
+        Value of the solution at (x, y) = (1, y).
+    """
+    return 0
+
+
+def g0(xy):
+    """Boundary condition at (x, y) = (x, 0).
+
+    Compute the value of the solution at (x, y) = (x, 0).
+
+    Parameters
+    ----------
+    xy : array-like of 2 float
+        Values for x and y, in that order.
+
+    Returns
+    -------
+    result : float
+        Value of the solution at (x, y) = (x, 0).
+    """
+    return 0
+
+
+def g1(xy):
+    """Boundary condition at (x, y) = (x, 1).
+
+    Compute the value of the solution at (x, y) = (x, 1).
+
+    Parameters
+    ----------
+    xy : array-like of 2 float
+        Values for x and y, in that order.
+
+    Returns
+    -------
+    result : float
+        Value of the solution at (x, y) = (x, 1).
+    """
[…]
