### 2021-12-18 02be72a
Final touch-ups to other software discussion.
