+++ paper.md
@@ -26,7 +26,7 @@ Neural networks have been shown to have the ability to solve differential equati
 
 The `nnde` package provides a pure-Python implementation of one of the earliest approaches to using neural networks to solve differential equations - the trial function method [@Lagaris:1998]. The `nnde` package was initially developed as a vehicle for understanding the internal workings of feedforward neural networks, without the constraints imposed by an existing neural network framework. It has since been enhanced to provide the capability to solve differential equations of scientific interest, such as the diffusion equation described here. The ultimate goal of the package is to provide the capability to solve systems of coupled partial differential equations, such as the equations of magnetohydrodynamics.
 
-Development of the `nnde` package began before the widespread adoption of modern neural network software frameworks, such as TensorFlow 2 (`https://tensorflow.org`) and PyTorch (`https://pytorch.org`). These frameworks provide autodifferentiation capability - by recording the sequence of mathematical operations performed in the forward pass through the network, the framework can automatically compute the gradients of the loss function with respect to each of the network parameters, as well as the network inputs. The latter capability is central to solving differential equations. The `nnde` package uses precomputed derivative functions for the components of the differential equation and the trial solution.
+Development of the `nnde` package began before the widespread adoption of modern neural network software frameworks. In the Python ecosystem, the most popular packages are TensorFlow (`https://tensorflow.org`) and PyTorch (`https://pytorch.org`). These frameworks are designed to be application-neutral - they can be used to develop neural networks with arbitrary architectures for arbitrary learning objectives. The primary advantages of these frameworks are autodifferentiation and distributed computing. By recording the sequence of mathematical operations performed in the forward pass through the network, autodifferentiation can automatically compute the gradients of the loss function with respect to each of the network parameters, as well as the network inputs. The latter capability is central to solving differential equations. Autodifferentiation also greatly reduces the volume of code that must be developed to solve a given problem. The distributed computing capability allows a network to take advantage of GPU-enabled hardware, and multiple compute nodes, to speed the calculation, with little or no code changes required. The `nnde` package uses a more direct method - precomputed derivative functions for the components of the differential equations of interest and the trial solution. This code is typically faster than TensorFlow or PyTorch, but requires more hand-crafted code to solve a given problem.
 
 The most commonly used methods for solving differential equations are the Finite Element Method (FEM) and Finite Difference Method (FDM). However, these methods can be difficult to parallelize due to the need for communication between computational elements at the boundaries of the allocated subgrids. These models can also  have large storage requirements for model outputs. The neural network method is straightforward to parallelize due to the independent characteristics of the computational nodes in each network layer. Additionally, the trained network solution is more compact than an FDM or FEM solution because storage of only the network weights and biases is required. The neural network solution is mesh-free and does not require interpolation to retrieve the solution at a non-grid point, as is the case with FDM or FEM. Once the network is trained, computing a solution at any spatial or temporal scale requires only a series of matrix multiplications, one per network layer. The trained solution is a sum of arbitrary differentiable basis functions, and therefore the trained solution is also differentiable, which is particularly useful when computing derived quantities such as gradients and fluxes. This approach has led to several different classes of methods for solving ODEs and PDEs with neural networks. The recent surge in interest in "physics-informed neural networks" [@Raissi:2019] is an indication of the dynamic nature of the field.
 
