### 2021-12-18 #3465 @elwinter
Hi Patrick. I apologize for the delay. I have reworked the paragraph
discussing TensorFlow and PyTorch to include a better comparison with nnde.
Please let me know what you think.

Thanks
Eric

On Mon, Dec 13, 2021 at 11:25 AM Patrick Diehl ***@***.***>
wrote:

> @elwinter <https://github.com/elwinter> could you please finish the paper
> this week? I really would like to get this paper out to take a new paper.
>
