### 2021-12-20 #3465 @diehlpk
@elwinter Thanks, the added text provided some benefit to the paper and I can recommend accepting it. Before, I can do that

- [ ] Please let me know the final version, including all changes during the JOSS review.
- [ ] Please upload this version to Zenodo or FigShare and post the DOI here. Note that the title and the author(s) have to match exactly with the papers.
