### 2022-01-19 #3465 @elwinter
Sorry for the delay.

Version 1.0
DOI 10.5281/zenodo.5879387

On Tue, Jan 18, 2022 at 11:41 AM Patrick Diehl ***@***.***>
wrote:

> @elwinter <https://github.com/elwinter> I really need the version and DOI
> to proceed. It would be bad to get your paper rejected at this stage.
>
