### 2022-02-07 #3465 @diehlpk
> OK. [10.5281/zenodo.5879387](https://doi.org/10.5281/zenodo.5879387) is the archive.

@elwinter If I click on the DOI, I see the following error:

![image](https://user-images.githubusercontent.com/3922435/152880712-266943bd-c852-4c74-9384-c84f14c52f84.png)
