### 2022-02-08 #3465 @elwinter
Ok checking that now.

Get Outlook for Android<https://aka.ms/ghei36>
