### 2022-02-08 #3465 @elwinter
OK try again when you have a chance. I had never used Zenodo before. My
file was there, and the URI was the same, but I apparently missed another
button that said "Publish".

Eric
