### 2022-02-08 #3465 @diehlpk
@elwinter One last thing

`nnde`: A Python package for solving differential equations using neural networks

The apostrophes around `nnde`: look strange. Would you mind removing them? You can just change the title without generating a new DOI.
