### 2022-02-08 #3465 @elwinter
Done. Sorry, missed that when I pasted in the title.

On Tue, Feb 8, 2022 at 3:19 PM Patrick Diehl ***@***.***>
wrote:

> @elwinter <https://github.com/elwinter> One last thing
>
> nnde: A Python package for solving differential equations using neural
> networks
>
> The apostrophes around nnde: look strange. Would you mind removing them?
> You can just change the title without generating a new DOI.
>
