### 2022-02-08 #3465 @elwinter
My thanks as well for your patience.  This was my first time for lot of this.

Take care
Eric

Get Outlook for Android<https://aka.ms/ghei36>
