### 2022-02-09 #3465 @kthyng
Hi @elwinter! A few items to finish up:

1. version and zenodo look good.
2. Does your co-author want to go by initials only?
3. In your paper, you shouldn't have an additional seciton underneath the affiliations stating your full listing.
4. Your reference section is repeated twice — not sure how that is happening!
