### 2022-02-14 #3465 @diehlpk
@elwinter I do not think anew DOI is necessary, just update the paper on GitHub. If I recall correctly, the paper.md from GitHub is used to generate the paper.
