### 2022-02-14 #3465 @elwinter
OK, but then the version published on Zenodo will differ from the version on GutHub. If that is OK, then I’ll make that change.

Eric

From: Patrick Diehl ***@***.***>
Sent: Monday, February 14, 2022 2:03 PM
Cc: Eric Winter ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: nnde: A Python package for solving differential equations using neural networks (#3465)

@elwinter <https://github.com/elwinter>  I do not think anew DOI is necessary, just update the paper on GitHub. If I recall correctly, the paper.md from GitHub is used to generate the paper.
