### 2022-02-14 #3465 @elwinter
OK, the updated version has been pushed to the master branch in the GitHub repository.

From: Patrick Diehl ***@***.***>
Sent: Monday, February 14, 2022 2:03 PM
Cc: Eric Winter ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: nnde: A Python package for solving differential equations using neural networks (#3465)

@elwinter <https://github.com/elwinter>  I do not think anew DOI is necessary, just update the paper on GitHub. If I recall correctly, the paper.md from GitHub is used to generate the paper.
