### 2022-02-14 #3465 @elwinter
Hi Kristen. I have pushed the updated version to my GitHub repo which removes the duplicate sections. And my co-author prefers to use leading initials and last name.

Thanks

Eric

From: Kristen Thyng ***@***.***>
Sent: Wednesday, February 9, 2022 6:30 PM
Cc: Eric Winter ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: nnde: A Python package for solving differential equations using neural networks (#3465)

Hi @elwinter <https://github.com/elwinter> ! A few items to finish up:

1.	version and zenodo look good.
2.	Does your co-author want to go by initials only?
3.	In your paper, you shouldn't have an additional seciton underneath the affiliations stating your full listing.
4.	Your reference section is repeated twice — not sure how that is happening!
