### 2022-02-14 #3465 @elwinter
Ah OK. Making the fixes now.

From: Daniel S. Katz ***@***.***>
Sent: Monday, February 14, 2022 4:06 PM
Cc: Eric Winter ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: nnde: A Python package for solving differential equations using neural networks (#3465)

The zenodo repository does not need to include the final version of the paper, just of the software
