### 2022-02-14 #3465 @diehlpk
> 2\. Does your co-author want to go by initials only?

@danielskatz The papers looks good now. The last thing is the question above.
