### 2022-02-19 #3465 @janmandel
Trying to install in google colab
!pip install PyAFBF
errors out with
`ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
albumentations 0.1.12 requires imgaug<0.2.7,>=0.2.5, but you have imgaug 0.2.9 which is incompatible.`
