### 2022-02-21 #3465 @janmandel
Sorry for the confusion. It was meant for another review. Thanks for
letting me know! This review system is getting the better of me.
Jan

On Sat, Feb 19, 2022 at 5:36 PM Eric Winter ***@***.***>
wrote:

> Hi Jan. I'm confused- my software was nnde, not PyAFBF. Is this a
> dependency collision of some kind?
>
> Eric
>
>
>
>
> Get Outlook for Android<https://aka.ms/ghei36>
> ________________________________
> From: Jan Mandel ***@***.***>
> Sent: Saturday, February 19, 2022 6:07:53 PM
> To: openjournals/joss-reviews ***@***.***>
> Cc: Eric Winter ***@***.***>; Mention ***@***.***>
> Subject: Re: [openjournals/joss-reviews] [REVIEW]: nnde: A Python package
> for solving differential equations using neural networks (#3465)
>
>
> Trying to install in google colab
> !pip install PyAFBF
> errors out with
> ERROR: pip's dependency resolver does not currently take into account all
> the packages that are installed. This behaviour is the source of the
> following dependency conflicts. albumentations 0.1.12 requires
> imgaug<0.2.7,>=0.2.5, but you have imgaug 0.2.9 which is incompatible.
>
