### 2022-02-21 #3465 @elwinter
Ahh, I thought that might be the case. No worries.

Take care
Eric

On Sun, Feb 20, 2022 at 11:31 PM Jan Mandel ***@***.***>
wrote:

> Sorry for the confusion. It was meant for another review. Thanks for
> letting me know! This review system is getting the better of me.
> Jan
>
> On Sat, Feb 19, 2022 at 5:36 PM Eric Winter ***@***.***>
> wrote:
>
> > Hi Jan. I'm confused- my software was nnde, not PyAFBF. Is this a
> > dependency collision of some kind?
> >
> > Eric
> >
> >
> >
> >
> > Get Outlook for Android<https://aka.ms/ghei36>
> > ________________________________
> > From: Jan Mandel ***@***.***>
> > Sent: Saturday, February 19, 2022 6:07:53 PM
> > To: openjournals/joss-reviews ***@***.***>
> > Cc: Eric Winter ***@***.***>; Mention ***@***.***>
> > Subject: Re: [openjournals/joss-reviews] [REVIEW]: nnde: A Python package
> > for solving differential equations using neural networks (#3465)
> >
> >
> > Trying to install in google colab
> > !pip install PyAFBF
> > errors out with
> > ERROR: pip's dependency resolver does not currently take into account all
> > the packages that are installed. This behaviour is the source of the
> > following dependency conflicts. albumentations 0.1.12 requires
> > imgaug<0.2.7,>=0.2.5, but you have imgaug 0.2.9 which is incompatible.
> >
> > —
> > Reply to this email directly, view it on GitHub<
> >
> https://github.com/openjournals/joss-reviews/issues/3465#issuecomment-1046121314
> >,
> > or unsubscribe<
> >
> https://github.com/notifications/unsubscribe-auth/ABSFWC5I26NPPXSPE44ZVRLU4APETANCNFSM5AC2IOEQ
> > >.
> > Triage notifications on the go with GitHub Mobile for iOS<
> >
> https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&mt=8&pt=524675
> >
> > or Android<
> >
> https://play.google.com/store/apps/details?id=com.github.android&referrer=utm_campaign%3Dnotification-email%26utm_medium%3Demail%26utm_source%3Dgithub
> > >.
> > You are receiving this because you were mentioned.Message ID:
> ***@***.***>
> >
> > —
> > Reply to this email directly, view it on GitHub
> > <
> https://github.com/openjournals/joss-reviews/issues/3465#issuecomment-1046132641
> >,
> > or unsubscribe
> > <
> https://github.com/notifications/unsubscribe-auth/ABPKZEFLKONHOYIAHXTMSVLU4AZR3ANCNFSM5AC2IOEQ
> >
> > .
> > You are receiving this because you commented.Message ID:
> > ***@***.***>
> >
>
